### 2023-10-16 #5955 @editorialbot
[PRE REVIEW]: HyperNetX: A Python package for modeling complex network data as hypergraphs
**Submitting author:** @brendapraggastis (<a href="http://orcid.org/0000-0003-1344-0497">Brenda Praggastis</a>)
**Repository:** https://github.com/pnnl/HyperNetX
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0.5
