+++ HNX_Paper/paper.bib
@@ -152,10 +152,8 @@
 title = {Phoenix: A Scalable Streaming Hypergraph Analysis Framework},
 author = {Kurte, Kuldeep and Imam, Neena and Hasan, S M Shamimul and Kannan, Ramakrishnan},
 doi = {10.1007/978-3-030-71704-9_1},
-url = {https://www.osti.gov/biblio/1830117}, journal = {},
+url = {https://www.osti.gov/biblio/1830117}, 
 issn = {2569--7072},
-number = ,
-volume = ,
 place = {United States},
 year = {2021},
 month = {10}
