### 2023-10-16 #5955 @brendapraggastis
It looks like it is compiling, at least I see a draft paper on github and it works when I run docker. I’m not sure what to fix. Is there somewhere else to look?

__________________________________
Brenda Praggastis, PhD
Senior Data Scientist and Mathematician
Math, Stats & Data Science
AI & Data Analytics

Pacific Northwest National Laboratory
1100 Dexter Ave N, Suite 500
Seattle, WA 98109 USA
(206) 528-3409 (voice only)

From: "Daniel S. Katz" ***@***.***>
Reply-To: openjournals/joss-reviews ***@***.***>
Date: Monday, October 16, 2023 at 12:46 PM
Cc: Brenda Praggastis ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: HyperNetX: A Python package for modeling complex network data as hypergraphs (Issue #5955)

Check twice before you click! This email originated from outside PNNL.

👋 @brendapraggastis<https://github.com/brendapraggastis> - note that your paper does not compile. Please follow the example paper<https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography> and note that you can click on the error above to find out more about it. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command @editorialbot check references to check again, and the command @editorialbot generate pdf after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
