### 2023-10-16 #5955 @danielskatz
> It looks like it is compiling, at least I see a draft paper on github and it works when I run docker. I’m not sure what to fix. Is there somewhere else to look?

See the error above re the Kurte, Kuldeep entry in the bibtex.  And use the commands I suggested - `@editorialbot check references` to make sure the bibtex is correct, then `@editorialbot generate pdf` to generate the paper, which won't work until the reference check is ok.

It seems that you are getting this figure out as I'm typing...
