### 2023-10-16 #5955 @danielskatz
And if `@editorialbot generate pdf` fails with no info now that the references can be generated, let me know and I can get help.
