### 2023-10-17 #5955 @brendapraggastis
@danielskatz how can I tell if the generate pdf command was successful? I just get a smiley face.
Also, do I need to do anything else to have the paper reviewed. I read somewhere that I should pick reviewers?
