### 2023-10-17 #5955 @danielskatz
@brendapraggastis - normally, you would see a response to the generate pdf command with a link to the PDF, like there is a response for the check references command.

Re the references, please do work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.
