### 2023-10-17 #5955 @danielskatz
@brendapraggastis - you can suggest reviewers, but it's the job of the editor to select and find reviewers, based in part on your suggestions. My job is to find an editor, but I won't do that until we have a clear paper and references, as I also do an initial check to make sure your submission is in scope, and I need to see the full submission (the software and the paper) to make an initial decision.
