### 2023-10-17 #5955 @danielskatz
Thanks @xuanxu - why don't get an error like we used to?
