+++ HNX_Paper/paper.bib
@@ -1,6 +1,7 @@
 
 @inproceedings{Joslyn2020DNS,
 	address = {Cham},
+  doi={10.1007/978-3-030-48478-1_1},
 	author = {Joslyn, Cliff A. and Aksoy, Sinan and Arendt, Dustin and Firoz, Jesun and Jenkins, Louis and Praggastis, Brenda and Purvine, Emilie and Zalewski, Marcin},
 	booktitle = {Algorithms and Models for the Web Graph},
 	editor = {Kami{\'{n}}ski, Bogumi{\l} and Pra{\l}at, Pawe{\l} and Szufel, Przemys{\l}aw},
@@ -12,6 +13,7 @@
 
 @article{Molnar2022Application,
   title={The application of directed hyper-graphs for analysis of models of information systems},
+  doi={10.3390/math10050759},
   author={Moln{\'a}r, B{\'a}lint and Bencz{\'u}r, Andr{\'a}s},
   journal={Mathematics},
   volume={10},
@@ -24,6 +26,7 @@
 /// Tony's work to quickly compute overlap sizes
 @inproceedings{Liu2021Parallel,
   title={Parallel algorithms for efficient computation of high-order line graphs of hypergraphs},
+  doi={10.1109/hipc53243.2021.00045},
   author={Liu, Xu T and Firoz, Jesun and Lumsdaine, Andrew and Joslyn, Cliff and Aksoy, Sinan and Praggastis, Brenda and Gebremedhin, Assefaw H},
   booktitle={2021 IEEE 28th International Conference on High Performance Computing, Data, and Analytics (HiPC)},
   pages={312--321},
@@ -33,6 +36,7 @@
 
 @article{Zhou2022Topological,
   title={Topological simplifications of hypergraphs},
+  doi={10.1109/tvcg.2022.3153895},
   author={Zhou, Youjia and Rathore, Archit and Purvine, Emilie and Wang, Bei},
   journal={IEEE Transactions on Visualization and Computer Graphics},
   year={2022},
@@ -41,6 +45,7 @@
 
 @article{Aksoy2020Hypernetwork,
   title={Hypernetwork science via high-order hypergraph walks},
+  doi={10.1140/epjds/s13688-020-00231-0},
   author={Aksoy, Sinan G and Joslyn, Cliff and Marrero, Carlos Ortiz and Praggastis, Brenda and Purvine, Emilie},
   journal={EPJ Data Science},
   volume={9},
@@ -52,6 +57,7 @@
 
 @article{Pickard2023HAT,
   title={HAT: Hypergraph analysis toolbox},
+  doi={10.1371/journal.pcbi.1011190},
   author={Pickard, Joshua and Chen, Can and Salman, Rahmy and Stansbury, Cooper and Kim, Sion and Surana, Amit and Bloch, Anthony and Rajapakse, Indika},
   journal={PLOS Computational Biology},
   volume={19},
@@ -63,6 +69,7 @@
 
 @misc{Berge1973Graphs,
   title={Graphs and {H}ypergraphs,(translated by {E}dward {M}inieka)},
+  doi={10.1016/s0924-6509(09)x7013-3},
   author={Berge, Claude},
   year={1973},
   publisher={American Elsevier, New York}
@@ -70,6 +77,7 @@
 
 @book{Berge1984Hypergraphs,
   title={Hypergraphs: combinatorics of finite sets},
+  doi={10.1016/s0924-6509(08)x7007-2},
   author={Berge, Claude},
   volume={45},
   year={1984},
@@ -78,6 +86,7 @@
 
 @article{Feng2021,
   title={Hypergraph models of biological networks to identify genes critical to pathogenic viral response},
[…]
