### 2023-10-17 796e27c
Modification of draw function to support drawing the star hypergraph
