+++ hypernetx/drawing/rubber_band.py
@@ -30,7 +30,7 @@ theta = np.linspace(0, 2 * np.pi, N_CONTROL_POINTS + 1)[:-1]
 cp = np.vstack((np.cos(theta), np.sin(theta))).T
 
 
-def layout_node_link(H, layout=nx.spring_layout, **kwargs):
+def layout_node_link(H, G=None, layout=nx.spring_layout, **kwargs):
     """
     Helper function to use a NetwrokX-like graph layout algorithm on a Hypergraph
 
@@ -41,6 +41,8 @@ def layout_node_link(H, layout=nx.spring_layout, **kwargs):
     ----------
     H: Hypergraph
         the entity to be drawn
+    G: Graph
+        an additional set of links to consider during the layout process
     layout: function
         the layout algorithm which accepts a NetworkX graph and keyword arguments
     kwargs: dict
@@ -51,7 +53,13 @@ def layout_node_link(H, layout=nx.spring_layout, **kwargs):
     dict
         mapping of node and edge positions to R^2
     """
-    return layout(H.bipartite(), **kwargs)
+
+    B = H.bipartite()
+    
+    if G is not None:
+        B.add_edges_from(G.edges())
+
+    return layout(B, **kwargs)
 
 
 def get_default_radius(H, pos):
@@ -82,7 +90,7 @@ def get_default_radius(H, pos):
     return 1
 
 
-def draw_hyper_edge_labels(H, polys, labels={}, ax=None, **kwargs):
+def draw_hyper_edge_labels(H, pos, polys, labels={}, edge_labels_on_edge=True, ax=None, **kwargs):
     """
     Draws a label on the hyper edge boundary.
 
@@ -113,22 +121,28 @@ def draw_hyper_edge_labels(H, polys, labels={}, ax=None, **kwargs):
     for edge, path, params in zip(H.edges, polys.get_paths(), params):
         s = labels.get(edge, edge)
 
-        # calculate the xy location of the annotation
-        # this is the midpoint of the pair of adjacent points the most distant
-        d = ((path.vertices[:-1] - path.vertices[1:]) ** 2).sum(axis=1)
-        i = d.argmax()
+        theta = 0
+        xy = pos[edge]
 
-        x1, x2 = path.vertices[i : i + 2]
-        x, y = x2 - x1
-        theta = 360 * np.arctan2(y, x) / (2 * np.pi)
-        theta = (theta + 360) % 360
+        if edge_labels_on_edge:
+            # calculate the xy location of the annotation
+            # this is the midpoint of the pair of adjacent points the most distant
+            d = ((path.vertices[:-1] - path.vertices[1:]) ** 2).sum(axis=1)
+            i = d.argmax()
 
-        while theta > 90:
-            theta -= 180
+            x1, x2 = path.vertices[i : i + 2]
+            x, y = x2 - x1
+            theta = 360 * np.arctan2(y, x) / (2 * np.pi)
+            theta = (theta + 360) % 360
+
+            while theta > 90:
+                theta -= 180
+
+            xy = (x1 + x2) / 2
 
         # the string is a comma separated list of the edge uid
         ax.annotate(
-            s, (x1 + x2) / 2, rotation=theta, ha="center", va="center", **params
+            s, xy, rotation=theta, ha="center", va="center", **params
         )
 
 
@@ -336,13 +350,17 @@ def draw(
     node_radius=None,
     edges_kwargs={},
     nodes_kwargs={},
+    edge_labels_on_edge=True,
     edge_labels={},
     edge_labels_kwargs={},
     node_labels={},
     node_labels_kwargs={},
     with_edge_labels=True,
     with_node_labels=True,
-    label_alpha=0.35,
[…]
