+++ HNX_Paper/paper.tex
@@ -1,87 +1,76 @@
----
-title: 'HyperNetX: A Python package for modeling complex network data as hypergraphs'
-tags:
-- Python
-- hypergraph
-- network science
-- simplicial-complexes
-- knowledge graph
-- simplicial-homology
-- s-linegraph
-- property hypergraph
-authors:
-- name: Brenda Praggastis
-  orcid: 0000-0003-1344-0497
-  affiliation: "1"
-- name: Sinan Aksoy
-  orcid: 0000-0002-3466-3334
-  affiliation: "1"
-- name: Dustin Arendt
-  orcid: 0000-0003-2466-199X
-  affiliation: "1"
-- name: Mark Bonicillo
-  affiliation: "1" 
-- name: Cliff Joslyn
-  orcid: 0000-0002-5923-5547
-  affiliation: "1"
-- name: Emilie Purvine
-  orcid: 0000-0003-2069-5594
-  affiliation: "1" 
-- name: Madelyn Shapiro
-  orcid: 0000-0002-2786-7056
-  affiliation: "1" 
-- name: Ji Young Yun
-  affiliation: "1"
-affiliations:
-- name: Pacific Northwest National Laboratory, USA
-  index: 1
-  
-date: 21 June 2023
-bibliography: paper.bib
+\documentclass{article}
+\usepackage[utf8]{inputenc}
+
+\title{HyperNetX: A Python package for modeling complex network data as hypergraphs}
+\author{Brenda Praggastis 
+    \and Sinan Aksoy 
+    \and Dustin Arendt 
+    \and Mark Bonicillo
+    \and Cliff Joslyn
+    \and Emilie Purvine
+    \and Madelyn Shapiro
+    \and Ji Young Yun}
+
+\date{June 2023}
+
+\usepackage{natbib}
+\usepackage{graphicx}
+
+\begin{document}
+
+\maketitle
  
----
- 
-# Summary
+\abstract{
 HyperNetX (HNX) is an open source Python library for the analysis and visualization of complex network data modeled as hypergraphs.
 Initially released in 2019, HNX facilitates exploratory data analysis of complex networks using algebraic topology, combinatorics, and generalized hypergraph and graph theoretical methods on structured data inputs.
 With its 2023 release, the library supports attaching metadata, numerical and categorical, to nodes (vertices) and  hyperedges, as well as to node-hyperedge pairings (incidences).
 HNX has a customizable Matplotlib-based visualization module as well as HypernetX-Widget, its JavaScript addon for interactive exploration and visualization of hypergraphs within Jupyter Notebooks. Both packages are available on GitHub and PyPI. With a growing community of users and collaborators, HNX has become a preeminent tool for hypergraph analysis. 
+}
 
-![HNX-Widget visualization of a Scene to Character mapping from the LesMis dataset [@knuth1993].](Figures/hnxexample.png){height="225pt"} 
+\begin{figure}[h!]
+    \centering
+    \includegraphics[height=225pt]{Figures/hnxexample.png}
+    \caption{HNX-Widget visualization of a Scene to Character mapping from the LesMis dataset \cite{knuth1993}.}
+    \label{fig:hnxexample}
+    \end{figure}
 
-
-# Statement of need
+\section{Statement of need}
 For more than a century, graph theory has provided powerful methods for studying network relationships among abstract entities.
-Since the early 2000's, software packages such as NetworkX [@hagberg2020] and Igraph [@csardi2006]  have
+Since the early 2000's, software packages such as NetworkX \cite{hagberg2020}and Igraph \cite{csardi2006} have
 made these theoretical tools available to data scientists for studying large data sets.
 Graphs represent pairwise interactions between entities, but for many network datasets this is a severe limitation.
[…]
