### 2023-10-18 0289273
HYP-356 Add deprecate warnings to certain ES methods
