### 2023-10-18 eb78a61
HYP-353 Add deprecation warnings for property column args
