+++ hypernetx/classes/entityset.py
@@ -34,6 +34,8 @@ class EntitySet:
         represents N-dimensional entity data (data table).
         Otherwise, represents 2-dimensional entity data (system of sets).
     data_cols : sequence of ints or strings, default=(0,1)
+    level1: str or int, default = 0
+    level2: str or int, default = 1
     data : numpy.ndarray, optional
         2D M x N ``ndarray`` of ``ints`` (data table);
         sparse representation of an N-dimensional incidence tensor with M nonzero cells.
@@ -73,9 +75,13 @@ class EntitySet:
         (order of columns does not matter; see Notes for an example).
         If doubly-nested dict,
         ``{item level: {item label: {property name: property value}}}``.
+    misc_props_col: str, default="properties"
+        Column names for miscellaneous properties, level index, and item name in
+        :attr:`properties`; see Notes for explanation.
     level_col: str, default="level"
     id_col : str,  default="id"
     cell_properties: sequence of int or str, pandas.DataFrame, or doubly-nested dict, optional
+    misc_cell_props_col: str, default="cell_properties"
 
     Notes
     -----
@@ -104,7 +110,10 @@ class EntitySet:
     all occurrences).
 
     The names of the Level (if provided) and ID columns must be specified by `level_col`
-    and `id_col`. All other columns will be considered explicit property types.
+    and `id_col`. `misc_props_col` can be used to specify the name of the column to be used
+    for miscellaneous properties; if no column by that name is found,
+    a new column will be created and populated with empty ``dicts``.
+    All other columns will be considered explicit property types.
     The order of the columns does not matter.
 
     This method assumes that there are no rows with the same (Level, ID);
@@ -129,12 +138,21 @@ class EntitySet:
         weights: Optional[Sequence[float] | float | int | str] = 1,
         aggregateby: Optional[str | dict] = "sum",
         properties: Optional[pd.DataFrame | dict[int, dict[T, dict[Any, Any]]]] = None,
+        misc_props_col: Optional[str] = None,
         level_col: str = "level",
         id_col: str = "id",
         cell_properties: Optional[
             Sequence[T] | pd.DataFrame | dict[T, dict[T, dict[Any, Any]]]
         ] = None,
+        misc_cell_props_col: Optional[str] = None,
     ):
+        if misc_props_col or misc_cell_props_col:
+            warnings.warn(
+                "misc_props_col and misc_cell_props_col will be deprecated; all public references to these "
+                "arguments will be removed in a future release.",
+                DeprecationWarning,
+            )
+
         self._uid = uid
         self._static = static
         self._state_dict = {}
