+++ hypernetx/classes/entityset.py
@@ -6,16 +6,19 @@ from ast import literal_eval
 from collections import OrderedDict, defaultdict
 from collections.abc import Hashable, Mapping, Sequence, Iterable
 from typing import Union, TypeVar, Optional, Any
+from typing_extensions import Self
 
 import numpy as np
 import pandas as pd
-from scipy.sparse import csr_matrix
+import scipy.sparse as sp
 
 from hypernetx.classes.helpers import (
     AttrList,
     assign_weights,
     remove_row_duplicates,
 )
+
+from hypernetx.utils.decorators import warn_to_be_deprecated
 
 T = TypeVar("T", bound=Union[str, int])
 
@@ -26,11 +29,13 @@ class EntitySet:
 
     Parameters
     ----------
-    entity : pandas.DataFrame, dict of lists or sets, list of lists or sets, optional
+    entity : pandas.DataFrame, dict of lists or sets, dict of dicts, list of lists or sets, optional
         If a ``DataFrame`` with N columns,
         represents N-dimensional entity data (data table).
         Otherwise, represents 2-dimensional entity data (system of sets).
-        TODO: Test for compatibility with list of Entities and update docs
+    data_cols : sequence of ints or strings, default=(0,1)
+    level1: str or int, default = 0
+    level2: str or int, default = 1
     data : numpy.ndarray, optional
         2D M x N ``ndarray`` of ``ints`` (data table);
         sparse representation of an N-dimensional incidence tensor with M nonzero cells.
@@ -45,7 +50,8 @@ class EntitySet:
         Ignored if `entity` is provided or `data` is not provided.
     uid : hashable, optional
         A unique identifier for the object
-    weights : str or sequence of float, optional
+    weight_col: string or int, default="cell_weights"
+    weights : sequence of float, float, int, str,  default=1
         User-specified cell weights corresponding to entity data.
         If sequence of ``floats`` and `entity` or `data` defines a data table,
             length must equal the number of rows.
@@ -54,11 +60,11 @@ class EntitySet:
         If ``str`` and `entity` is a ``DataFrame``,
             must be the name of a column in `entity`.
         Otherwise, weight for all cells is assumed to be 1.
-    aggregateby : {'sum', 'last', count', 'mean','median', max', 'min', 'first', None}
+    aggregateby : {'sum', 'last', count', 'mean','median', max', 'min', 'first', None}, default="sum"
         Name of function to use for aggregating cell weights of duplicate rows when
-        `entity` or `data` defines a data table, default is "sum".
+        `entity` or `data` defines a data table.
         If None, duplicate rows will be dropped without aggregating cell weights.
-        Effectively ignored if `entity` defines a system of sets.
+        Ignored if `entity` defines a system of sets.
     properties : pandas.DataFrame or doubly-nested dict, optional
         User-specified properties to be assigned to individual items in the data, i.e.,
         cell entries in a data table; sets or set elements in a system of sets.
@@ -66,12 +72,16 @@ class EntitySet:
         If ``DataFrame``, each row gives
         ``[optional item level, item label, optional named properties,
         {property name: property value}]``
-        (order of columns does not matter; see note for an example).
+        (order of columns does not matter; see Notes for an example).
         If doubly-nested dict,
         ``{item level: {item label: {property name: property value}}}``.
-    misc_props_col, level_col, id_col : str, default="properties", "level, "id"
+    misc_props_col: str, default="properties"
         Column names for miscellaneous properties, level index, and item name in
         :attr:`properties`; see Notes for explanation.
+    level_col: str, default="level"
[…]
