### 2023-10-25 a249417
HYP-177 Reorg entityset tests
