+++ hypernetx/classes/tests/conftest.py
@@ -42,8 +42,8 @@ class SevenBySix:
         )
         self.labels = OrderedDict(
             [
-                ("edges", ["P", "R", "S", "L", "O", "I"]),
-                ("nodes", ["A", "C", "E", "K", "T1", "T2", "V"]),
+                ("edges", [p, r, s, l, o, i]),
+                ("nodes", [a, c, e, k, t1, t2, v]),
             ]
         )
 
@@ -51,18 +51,18 @@ class SevenBySix:
             [
                 [0, 0],
                 [0, 1],
-                [0, 2],
+                [0, 3],
+                [1, 0],
                 [1, 2],
-                [1, 3],
                 [2, 0],
-                [2, 2],
-                [2, 4],
+                [2, 3],
                 [2, 5],
+                [2, 6],
                 [3, 1],
-                [3, 3],
+                [3, 2],
+                [4, 4],
                 [4, 5],
-                [4, 6],
-                [5, 0],
+                [5, 3],
                 [5, 5],
             ]
         )
