+++ hypernetx/classes/tests/test_entityset_sbs_data.py
@@ -1,619 +1,0 @@
-import numpy as np
-import pandas as pd
-import pytest
-
-from pytest_lazyfixture import lazy_fixture
-
-from hypernetx.classes import EntitySet
-
-
-@pytest.mark.parametrize(
-    "entity, data, data_cols, labels",
-    [
-        (lazy_fixture("sbs_dataframe"), None, (0, 1), None),
-        (lazy_fixture("sbs_dict"), None, (0, 1), None),
-        (lazy_fixture("sbs_dict"), None, ["edges", "nodes"], None),
-        # (None, lazy_fixture("sbs_data"), (0, 1), lazy_fixture("sbs_labels")),
-    ],
-)
-class TestEntitySetUseCasesOnSBS:
-    # Tests on different use cases for combination of the following params: entity, data, data_cols, labels
-
-    def test_size(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert es.size() == len(sbs.edgedict)
-
-    # check all the EntitySet properties
-    def test_isstatic(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert es.isstatic
-
-    def test_uid(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert es.uid is None
-
-    def test_empty(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert not es.empty
-
-    def test_uidset(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert es.uidset == {"I", "R", "S", "P", "O", "L"}
-
-    def test_dimsize(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert es.dimsize == 2
-
-    def test_elements(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        assert len(es.elements) == 6
-        expected_elements = {
-            "I": ["K", "T2"],
-            "L": ["E", "C"],
-            "O": ["T1", "T2"],
-            "P": ["C", "K", "A"],
-            "R": ["E", "A"],
-            "S": ["K", "V", "A", "T2"],
-        }
-        for expected_edge, expected_nodes in expected_elements.items():
-            assert expected_edge in es.elements
-            assert es.elements[expected_edge].sort() == expected_nodes.sort()
-
-    def test_incident_dict(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
-        expected_incident_dict = {
-            "I": ["K", "T2"],
-            "L": ["E", "C"],
-            "O": ["T1", "T2"],
-            "P": ["C", "K", "A"],
-            "R": ["E", "A"],
-            "S": ["K", "V", "A", "T2"],
-        }
-        for expected_edge, expected_nodes in expected_incident_dict.items():
-            assert expected_edge in es.incidence_dict
-            assert es.incidence_dict[expected_edge].sort() == expected_nodes.sort()
-        assert isinstance(es.incidence_dict["I"], list)
-        assert "I" in es
-        assert "K" in es
-
-    def test_children(self, entity, data, data_cols, labels, sbs):
[…]
