### 2023-11-01 #5955 @szhorvat
@danielskatz This is within my expertise and I've seen this package before. I think a _proper_ review will take time. This is not a one-afternoon job.  I would not be able to get started before Nov 13 (when teaching is over), and I might take a while to finish ... probably several weeks. If you'd like me to go ahead under these conditions, I think I'd enjoy the review.
