### 2023-11-01 e1b6d1b
Cleanup makefile
