### 2023-11-01 74a9977
Update tests for soon to be deprecated translate methods
