### 2023-11-03 2656a06
Fix github workflow for documentation
