+++ .github/workflows/documentation.yml
@@ -10,7 +10,6 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install sphinx sphinx_rtd_theme
           pip install .'[documentation]'
       - name: Sphinx build
         run: |
