### 2023-11-03 3dbf148
Remove unnecessary dependencies for documentation
