+++ setup.cfg
@@ -97,7 +97,6 @@ widget =
 	jupyter-nbextensions-configurator>=0.6.2
 documentation =
 	sphinx<7
-	nb2plots>=0.7.2
 	sphinx-rtd-theme>=1.2.1
 	sphinx-autobuild>=2021.3.14
 	sphinx-copybutton>=0.5.1
