### 2023-11-03 a29cf8d
Remove unnecessary dependencies for documentation
