+++ hypernetx/algorithms/hypergraph_modularity.py
@@ -16,8 +16,7 @@ from collections import Counter
 import numpy as np
 import pandas as pd
 import itertools
-from scipy.special import comb
-from scipy.stats import binom 
+from scipy.stats import binom
 
 try:
     import igraph as ig
@@ -76,6 +75,7 @@ def part2dict(A):
 
 
 ################################################################################
+
 
 def linear(d, c):
     """
@@ -172,60 +172,63 @@ def modularity(HG, A, wdc=linear):
       The modularity function for partition A on HG
     """
     ## all same edge weights?
-    uniq = (len(Counter(HG.edges.properties['weight']))==1)
-    
+    uniq = len(Counter(HG.edges.properties["weight"])) == 1
+
     ## Edge Contribution
     HG_id = HG.incidence_dict
     d = part2dict(A)
-    L = [ [d[i] for i in HG_id[x]] for x in HG_id ]
+    L = [[d[i] for i in HG_id[x]] for x in HG_id]
 
     ## all same weight
     if uniq:
-        _ctr = Counter([ (Counter(l).most_common(1)[0][1],len(l)) for l in L])
-        EC = sum([wdc(k[1],k[0])*_ctr[k] for k in _ctr.keys() if k[0] > k[1]/2])
+        _ctr = Counter([(Counter(l).most_common(1)[0][1], len(l)) for l in L])
+        EC = sum([wdc(k[1], k[0]) * _ctr[k] for k in _ctr.keys() if k[0] > k[1] / 2])
     else:
-        _keys = [ (Counter(l).most_common(1)[0][1],len(l)) for l in L]
-        _vals = list(HG.edge_props['weight']) ## Thanks Brenda!!
-        _df = pd.DataFrame(zip(_keys,_vals), columns=['key','val'])
-        _df = _df.groupby(by='key').sum()
-        EC = sum([ wdc(k[1],k[0])*v[0] for (k,v) in _df.iterrows() if k[0]>k[1]/2 ])
-        
+        _keys = [(Counter(l).most_common(1)[0][1], len(l)) for l in L]
+        _vals = list(HG.edge_props["weight"])  ## Thanks Brenda!!
+        _df = pd.DataFrame(zip(_keys, _vals), columns=["key", "val"])
+        _df = _df.groupby(by="key").sum()
+        EC = sum(
+            [wdc(k[1], k[0]) * v[0] for (k, v) in _df.iterrows() if k[0] > k[1] / 2]
+        )
+
     ## Degree Tax
-    if uniq:        
+    if uniq:
         VolA = [sum([HG.degree(i) for i in k]) for k in A]
         Ctr = Counter([HG.size(i) for i in HG.edges])
 
     else:
         ## this is the bottleneck
-        VolA = np.repeat(0,1+np.max(list(d.values())))
+        VolA = np.repeat(0, 1 + np.max(list(d.values())))
         m = np.max([HG.size(i) for i in HG.edges])
-        Ctr = np.repeat(0,1+m)
+        Ctr = np.repeat(0, 1 + m)
         S = 0
         for e in HG.edges:
             w = HG.edges[e].weight
-            Ctr[HG.size(e)] += w  
+            Ctr[HG.size(e)] += w
             S += w
             for v in HG.edges[e]:
-                VolA[d[v]] += w 
-                
+                VolA[d[v]] += w
+
     VolV = np.sum(VolA)
-    VolA = [x/VolV for x in VolA]
+    VolA = [x / VolV for x in VolA]
     DT = 0
-    
-    if uniq:        
+
+    if uniq:
         for d in Ctr.keys():
             Cnt = Ctr[d]
-            for c in np.arange(int(np.floor(d/2+1)),d+1):
[…]
