### 2023-11-15 pnnl/HyperNetX#126 @IvanIsCoding
[joss] Improve class signatures in documentation
Currently, class signatures in the documentation are hard to read due to type annotations crowding:

![Captura de tela de 2023-11-14 22-48-50](https://github.com/pnnl/HyperNetX/assets/8753214/1be24750-36b2-48df-bfd2-fe23a974fb97)

In the example above, there are so many type annotations that it is hard to find the second paramater `data`

**Describe the solution you'd like**

Because the parameters are already described in the docstring, I belive that hiding the type annotations would lead to a better documentation. Sometimes less is more. The parameters for the class constructors should be obvious.

**Additional context**

https://github.com/openjournals/joss-reviews/issues/6016, I do not consider this issue blocking for acceptance. But it would be a good improvement
