### 2023-11-15 #6016 @IvanIsCoding
I have reviewed the software package and preliminary I believe it would be a great addition to the papers published at the Journal.

## Software

With regards to the software, I do not have objections. I believe HyperNetX is a significant scholarly contribution and that it is as mature as if not more mature than other packages published in this journal.

From a Software Engineering point of view, I also found the that the test suite of the package meets the standards for publication. I was able to install the software and run the tests without issues. I was also able to generate a code coverage report. There are some warnings when running the tests and the coverage for the Drawing module is on the lower side. But overall the software is tested and Continuous Integration is in place to detect regressions, which is nice.

## Documentation

For the documentation I focused more on the provided Jupyter Notebooks and [A Gentle Introduction to Hypergraph Mathematics](https://pnnl.github.io/HyperNetX/hypergraph101.html#a-gentle-introduction-to-hypergraph-mathematics) provided in the docs. There are also documentations for every function and class, but I admit I haven't scrutinized each method and looked at the greater picture.

As a reviewer with a greater emphasis on graphs than hypegraphs, I found that the "gentle" introduction answered most of the questions for the statement of need and defined the terminology for the package.

The Jupyter Notebooks available in Colab were also a pleasant surprise. I think they exemplify how to use the package and highlight the visualization features that come with HyperNetX.

## Paper

This is the only section where I think there is feedback to give to the authors from my side. Namely:
* The "HNX-Widget: Interactive Diagrams of Hypergraphs" is my least favourite section of the paper. I do not believe it aggregates much given the word limit recommended by JOSS. I suggest using an image with a detailed caption to convey the message better
* [The authors should cite `pandas` following the guidelines of the package](https://pandas.pydata.org/about/citing.html)
* [The authors should cite `NetworkX` following the guidelines of the package](https://networkx.org/documentation/stable/index.html#citing)
