### 2023-11-15 pnnl/HyperNetX#126 @bonicim
@IvanIsCoding Thanks for your feedback. We are currently overhauling our documentation to be more user-friendly and will address your comments regarding how the type annotations are obscuring the constructor and parameters.
