### 2023-11-18 937c226
HYP-360 Improve code quality tooling; update Makefile
