+++ setup.cfg
@@ -72,6 +72,14 @@ install_requires =
 [options.extras_require]
 releases =
 	commitizen>=3.2.1
+lint =
+	pylint>=3.0.2
+    pylint-exit>=1.2.0
+	mypy>=1.7.0
+	flake8>=6.1.0
+	pre-commit>=3.2.2
+format =
+	black>=23.3.0
 testing =
 	pytest>=7.2.2
 	pytest-cov>=4.1.0
@@ -80,10 +88,6 @@ testing =
 	pytest-env
 	tox>=4.4.11
 	nbmake>=1.4.1
-	pre-commit>=3.2.2
-    pylint>=2.17.2
-    pylint-exit>=1.2.0
-    black>=23.3.0
 	celluloid>=0.2.0
 	igraph>=0.10.4
 tutorials =
