### 2023-11-22 bc835ae
updates as per JOSS reviewer recommendations.
