+++ HNX_Paper/paper.md
@@ -1,0 +1,120 @@
+---
+title: 'HyperNetX: A Python package for modeling complex network data as hypergraphs'
+tags:
+- Python
+- hypergraph
+- network science
+- simplicial-complexes
+- knowledge graph
+- simplicial-homology
+- s-linegraph
+- property hypergraph
+authors:
+- name: Brenda Praggastis
+  orcid: 0000-0003-1344-0497
+  affiliation: "1"
+- name: Sinan Aksoy
+  orcid: 0000-0000-0000-0000
+  affiliation: "1"
+- name: Dustin Arendt
+  orcid: 0000-0000-0000-0000
+  affiliation: "1"
+- name: Mark Bonicillo
+  orcid: 0000-0000-0000-0000
+  affiliation: "1" 
+- name: Cliff Joslyn
+  orcid: 0000-0000-0000-0000
+  affiliation: "1"
+- name: Emilie Purvine
+  orcid: 0000-0000-0000-0000
+  affiliation: "1" 
+- name: Madelyn Shapiro
+  orcid: 0000-0000-0000-0000
+  affiliation: "1" 
+- name: Ji Young Yun
+  orcid: 0000-0000-0000-0000
+  affiliation: "1"
+affiliations:
+- name: Pacific Northwest National Laboratory, USA
+  index: 1
+  
+date: 21 June 2023
+bibliography: paper.bib
+ 
+---
+ 
+# Summary
+HyperNetX (HNX) is an open source Python library for the analysis and visualization of complex network data modeled as hypergraphs.
+Initially released in 2019, HNX facilitates exploratory data analysis of complex networks using algebraic topology, combinatorics, and generalized hypergraph and graph theoretical methods on structured data inputs.
+With its 2023 release, the library supports attaching metadata, numerical and categorical, to nodes (vertices) and  hyperedges, as well as to node-hyperedge pairings (incidences).
+HNX has a customizable Matplotlib-based visualization module as well as HypernetX-Widget, its JavaScript addon for interactive exploration and visualization of hypergraphs within Jupyter Notebooks. Both packages are available on GitHub and PyPI. With a growing community of users and collaborators, HNX has become a preeminent tool for hypergraph analysis. 
+
+![HNX-Widget is an add-on for the Jupyter Notebook
+computational environment, enabling users to view and interactively
+explore hypergraphs.
+The main features of the tool are: 1) adjustable layout 2) advanced
+selection and 3) visual encoding of node and edge properties.
+Metadata may be attached to the tool by providing tabular data via two optional data frames indexed by node and hyperedge identifiers. Above is an HNX-Widget visualization of a Scene to Character mapping from the LesMis dataset [@knuth1993].](Figures/hnxexample.png){height="225pt"} 
+
+
+# Statement of need
+For more than a century, graph theory has provided powerful methods for studying network relationships among abstract entities.
+Since the early 2000's, software packages such as NetworkX [@SciPyProceedings_11] and Igraph [@csardi2006]  have
+made these theoretical tools available to data scientists for studying large data sets.
+Graphs represent pairwise interactions between entities, but for many network datasets this is a severe limitation.
+In 1973, hypergraphs were introduced by Claude Berge [@Berge1973Graphs] as a strict generalization of graphs: a hyperedge in a hypergraph can contain any number of nodes, including 1, 2, or more.
+Hypergraphs have been used to model complex network datasets in
+areas such as the biological sciences, information systems, the power grid, and cyber security.
[…]
