+++ HNX_Paper/paper.bib
@@ -169,12 +169,6 @@ month = {10}
   publisher={AcM Press New York}
 }
 
-@article{hagberg2020,
-  title={Networkx: Network analysis with python},
-  author={Hagberg, Aric and Conway, Drew},
-  journal={URL: https://networkx. github. io},
-  year={2020}
-}
 
 @article{csardi2006,
   title={The igraph software package for complex network research},
@@ -227,3 +221,33 @@ month = {10}
 	bdsk-url-1 = {https://www.biorxiv.org/content/early/2023/10/02/2023.09.29.560191},
 	bdsk-url-2 = {https://doi.org/10.1101/2023.09.29.560191}}
 
+@software{reback2020pandas,
+    author       = {{The pandas development team}},
+    title        = {{pandas-dev/pandas: Pandas}},
+    month        = feb,
+    year         = 2020,
+    publisher    = {Zenodo},
+    version      = {latest},
+    doi          = {10.5281/zenodo.3509134},
+    url          = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@InProceedings{ mckinney-proc-scipy-2010,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
+}
+
+@InProceedings{SciPyProceedings_11,
+  author =       {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
+  title =        {Exploring Network Structure, Dynamics, and Function using NetworkX},
+  booktitle =   {Proceedings of the 7th Python in Science Conference},
+  pages =     {11 - 15},
+  address = {Pasadena, CA USA},
+  year =      {2008},
+  editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
+}
