### 2023-11-22 #6016 @brendapraggastis
@IvanIsCoding  Thanks for the detailed review. I've incorporated your suggesttions in the paper in https://github.com/pnnl/HyperNetX/tree/paper/HNX_Paper.
