### 2023-11-30 pnnl/HyperNetX#127 @drj11
The LICENSE is not clearly OSI compliant
**Describe the bug**

LICENSE is provided in `LICENSE.rst` but it not clearly OSI compliant.

**Additional context**

I am reviewing this package for JOSS https://github.com/openjournals/joss-reviews/issues/6016#issuecomment-1792288685

It looks like the license is intended to be the BSD 3-clause license (at OSI here: https://opensource.org/license/bsd-3-clause/ ) but:

- there is a typo at the end of the 2nd clause which currently reads "and/or other" but should read "and/or other materials provided with the distribution."
- the third clause has been edited in a way as to make it unclear if it still matches the BSD 3-clause license. In particular "Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software" has been changed to "neither the name Battelle Memorial Institute or Battelle may be used in any form whatsoever without the express written consent of Battelle.".

The modification to clause 3 appears to be a substantial reduction in the grant.

I suggest that you either:

- use the BSD 3-clause license directly; or,
- provide enough material to reviewers to assure us that this license is in fact "the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license"
