### 2023-12-04 pnnl/HyperNetX#127 @danielskatz
For JOSS to review the work, it has to have an OSI-compliant license, which means that the license can't be a variation of an OSI license, but actually an OSI license.
