### 2023-12-04 fccef29
Run linter on LICENSE.rst
