+++ .cz.toml
@@ -1,6 +1,6 @@
 [tool.commitizen]
 name = "cz_conventional_commits"
-version = "2.1.1"
+version = "2.1.2"
 version_files = [
     "setup.py",
     "docs/source/conf.py",
