### 2023-12-06 #6016 @drj11
I have completed my review. While i have no poked into every corner of functionality described in the paper, i have poked around enough to be satisfied that the functionality of HyperNetX is substantially as described in its documentation. I find no further reason to delay publication, and recommend that this paper be published by JOSS.

I find the subject to be quite novel and interesting (particularly coming from a more trad graph theory background), and i liked the inclusion of Les Misérables in the tutorials. The package as a whole is made to a high standard, and integrates well with the established Python ecosystems. It is a credit to both the Python and the scientific programming community. I commend the authors for their efforts.

My only real substantial comment, and it is relatively minor, is that i would spent some time on your very long `pylint` and my own particular bugbear: `bare-except`.
