### 2023-12-08 59ddab7
HYP-364 Fix doctrings
