+++ setup.py
@@ -1,5 +1,5 @@
 from setuptools import setup
 
-__version__ = "2.1.2"
+__version__ = "2.1.3"
 
 setup(version=__version__)
