### 2023-12-20 4383c37
updated empty hypergraph to have dataframes for properties
