+++ hypernetx/classes/tests/test_entityset.py
@@ -1,371 +1,0 @@
-import numpy as np
-import pytest
-
-from collections.abc import Iterable
-from collections import UserList
-from hypernetx.classes import EntitySet
-from hypernetx.classes.entityset import restrict_to_two_columns
-
-from pandas import DataFrame, Series
-
-def test_empty_entityset():
-
-    es = EntitySet()
-    assert es.empty
-    assert len(es.elements) == 0
-    assert es.elements == {}
-    assert es.dimsize == 0
-
-
-def test_entityset_from_dataframe():
-    data_dict = {
-        1: ["A", "D"],
-        2: ["A", "C", "D"],
-        3: ["D"],
-        4: ["A", "B"],
-        5: ["B", "C"],
-    }
-
-    all_edge_pairs = Series(data_dict).explode()
-
-    entity = DataFrame(
-        {"edges": all_edge_pairs.index.to_list(), "nodes": all_edge_pairs.values}
-    )
-
-    es = EntitySet(entity=entity)
-
-    assert not es.empty
-    assert len(es.elements) == 5
-    assert es.dimsize == 2
-    assert es.uid is None
-
-
-class TestEntitySetOnSevenBySixDataset:
-    # Tests on different inputs for entity and data
-    def test_entityset_from_dictionary(self, sbs):
-        ent = EntitySet(entity=sbs.edgedict)
-        assert len(ent.elements) == 6
-
-    def test_entityset_from_ndarray_sbs(self, sbs):
-        ent_sbs = EntitySet(data=np.asarray(sbs.data), labels=sbs.labels)
-
-        assert ent_sbs.size() == 6
-        assert len(ent_sbs.uidset) == 6
-        assert len(ent_sbs.children) == 7
-        assert isinstance(ent_sbs.incidence_dict["I"], list)
-        assert "I" in ent_sbs
-        assert "K" in ent_sbs
-
-    # Tests for properties
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_cell_properties(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_cell_weights(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_children(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_data(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_dataframe(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_dimensions(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_dimsize(self):
-        pass
-
-    def test_dimensions_equal_dimsize(self, sbs):
-        ent_sbs = EntitySet(data=np.asarray(sbs.data), labels=sbs.labels)
-        assert ent_sbs.dimsize == len(ent_sbs.dimensions)
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_elements(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_empty(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_incidence_dict(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_isstatic(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_labels(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_memberships(self):
-        pass
-
-    @pytest.mark.skip(reason="TODO: implement")
-    def test_properties(self):
-        pass
-
[…]
