### 2024-01-02 ace1144
HYP-365 Silence warnings in tutorials
