+++ setup.cfg
@@ -95,6 +95,7 @@ tutorials =
 	igraph>=0.10.4
 	partition-igraph>=0.0.6
 	celluloid>=0.2.0
+	shutup>=0.2.0
 widget =
 	hnxwidget>=0.1.1b3
 	jupyter-contrib-nbextensions>=0.7.0
