### 2024-01-03 #6016 @bonicim
@IvanIsCoding I have addressed the issues that your raised and merged in the changes in both the `paper` and `master` branch; the changes can be referenced via the commit message. Any commit message prefixed with `HYP-364` refers to the requested changes for this review. @szhorvat The [`paper` branch](https://github.com/pnnl/HyperNetX/tree/paper) has been updated and ready for review. Please let me know if you need anything else or have any questions.
