### 2024-01-16 6a0c087
HYP-367 Add stubs for PropertyStore
