### 2024-01-16 23471bf
Continued stubbing out incidenceStore
