### 2024-01-16 3bbee81
Adding more detailed stubs
