### 2024-01-17 1128fe2
HYP-367 Add factory method stubs for PropertyStoreFactory; update tests; add methods to HarryPotter toy
