### 2024-01-17 1ec92ec
Adding comments at top of each method in hypergraph.py and entityset.py for where new methods will belong in stores.
