+++ hypernetx/classes/entityset.py
@@ -264,6 +264,7 @@ class EntitySet:
 
     @property
     def cell_properties(self) -> Optional[pd.DataFrame]:
+        # method on incidence stores
         """Properties assigned to cells of the incidence matrix
 
         Returns
@@ -275,6 +276,7 @@ class EntitySet:
 
     @property
     def data(self) -> np.ndarray:
+        # method on incidence store
         """Sparse representation of the data table as an incidence tensor
 
         This can also be thought of as an encoding of `dataframe`, where items in each column of
@@ -306,6 +308,7 @@ class EntitySet:
 
     @property
     def labels(self) -> dict[str, list]:
+        # method on incidence store
         """Labels of all items in each column of the underlying data table
 
         Returns
@@ -331,6 +334,7 @@ class EntitySet:
 
     @property
     def cell_weights(self) -> dict[str, tuple[T]]:
+        # method on incidence store
         """Cell weights corresponding to each row of the underlying data table
 
         Returns
@@ -351,6 +355,7 @@ class EntitySet:
 
     @property
     def dimensions(self) -> tuple[int]:
+        # method on incidence store
         """Dimensions of data i.e., the number of distinct items in each level (column) of the underlying data table
 
         Returns
@@ -371,6 +376,7 @@ class EntitySet:
 
     @property
     def dimsize(self) -> int:
+        # method on incidence store
         """Number of levels (columns) in the underlying data table
 
         Returns
@@ -382,6 +388,7 @@ class EntitySet:
 
     @property
     def properties(self) -> pd.DataFrame:
+        # method using incidence store and property stores
         """Properties assigned to items in the underlying data table
 
         Returns
@@ -393,6 +400,7 @@ class EntitySet:
 
     @property
     def uid(self) -> Hashable:
+        #?
         """User-defined unique identifier for the `Entity`
 
         Returns
@@ -403,6 +411,7 @@ class EntitySet:
 
     @property
     def uidset(self) -> set:
+        #?
         """Labels of all items in level 0 (first column) of the underlying data table
 
         Returns
@@ -419,6 +428,7 @@ class EntitySet:
 
     @property
     def children(self) -> set:
+        #method on incidence store?
         """Labels of all items in level 1 (second column) of the underlying data table
 
         Returns
@@ -434,6 +444,7 @@ class EntitySet:
         return self.uidset_by_level(1)
 
     def uidset_by_level(self, level: int) -> set:
+        #method on incidence store?
         """Labels of all items in a particular level (column) of the underlying data table
 
         Parameters
@@ -456,6 +467,7 @@ class EntitySet:
         return self.uidset_by_column(col)
 
     def uidset_by_column(self, column: Hashable) -> set:
+        #method on incidence store?
         # TODO: This threw an error when trying it on the harry potter dataset,
         # when trying 0, or 1 for column. I'm not sure how this should be used
         """Labels of all items in a particular column (level) of the underlying data table
@@ -487,6 +499,7 @@ class EntitySet:
 
     @property
     def elements(self) -> dict[Any, AttrList]:
+        #method on incidence store
         """System of sets representation of the first two levels (columns) of the underlying data table
 
         Each item in level 0 (first column) defines a set containing all the level 1
@@ -515,6 +528,7 @@ class EntitySet:
 
     @property
     def incidence_dict(self) -> dict[T, list[T]]:
+        #method on incidence store
         """System of sets representation of the first two levels (columns) of the underlying data table
 
         Returns
@@ -531,6 +545,7 @@ class EntitySet:
 
     @property
     def memberships(self) -> dict[Any, AttrList]:
+        #method on incidence store
[…]
