+++ hypernetx/classes/hypergraph.py
@@ -574,6 +574,7 @@ class Hypergraph:
 
     @property
     def edges(self):
+        # returns property store on edges
         """
         Object associated with self._edges.
 
@@ -585,6 +586,7 @@ class Hypergraph:
 
     @property
     def nodes(self):
+        # returns property store on nodes
         """
         Object associated with self._nodes.
 
@@ -596,6 +598,7 @@ class Hypergraph:
 
     @property
     def dataframe(self):
+        # returns incidence store 
         """Returns dataframe of incidence pairs and their properties.
 
         Returns
@@ -606,16 +609,18 @@ class Hypergraph:
 
     @property
     def properties(self):
+        # method on hypergraph class that concatenates the property stores on edges and nodes
         """Returns dataframe of edge and node properties.
 
         Returns
         -------
         pd.DataFrame
         """
-        return self.E.properties
+        return self.E.properties 
 
     @property
     def edge_props(self):
+        # Method on property store on edges
         """Dataframe of edge properties
         indexed on edge ids
 
@@ -627,6 +632,7 @@ class Hypergraph:
 
     @property
     def node_props(self):
+        # property store on nodes
         """Dataframe of node properties
         indexed on node ids
 
@@ -638,6 +644,7 @@ class Hypergraph:
 
     @property
     def incidence_dict(self):
+        # method on incidence store
         """
         Dictionary keyed by edge uids with values the uids of nodes in each
         edge
@@ -651,6 +658,7 @@ class Hypergraph:
 
     @property
     def shape(self):
+        # method on incidence store
         """
         (number of nodes, number of edges)
 
@@ -662,6 +670,7 @@ class Hypergraph:
         return len(self._nodes.elements), len(self._edges.elements)
 
     def __str__(self):
+        # method on incidence store
         """
         String representation of hypergraph
 
@@ -673,6 +682,7 @@ class Hypergraph:
         return f"{self.name}, <class 'hypernetx.classes.hypergraph.Hypergraph'>"
 
     def __repr__(self):
+        # method on incidence store
         """
         String representation of hypergraph
 
@@ -684,6 +694,7 @@ class Hypergraph:
         return f"{self.name}, hypernetx.classes.hypergraph.Hypergraph"
 
     def __len__(self):
+        # method on property store of nodes
         """
         Number of nodes
 
@@ -695,6 +706,7 @@ class Hypergraph:
         return len(self._nodes)
 
     def __iter__(self):
+        # method on property store of nodes
         """
         Iterate over the nodes of the hypergraph
 
@@ -706,6 +718,7 @@ class Hypergraph:
         return iter(self.nodes)
 
     def __contains__(self, item):
+        # method on property store of nodes
         """
         Returns boolean indicating if item is in self.nodes
 
@@ -717,6 +730,7 @@ class Hypergraph:
         return item in self.nodes
 
     def __getitem__(self, node):
+        # method on incidence store
         """
         Returns the neighbors of node
 
@@ -735,6 +749,7 @@ class Hypergraph:
     def get_cell_properties(
         self, edge: str, node: str, prop_name: Optional[str] = None
     ) -> Any | dict[str, Any]:
+        # method on incidence store
         """Get cell properties on a specified edge and node
 
         Parameters
@@ -758,6 +773,7 @@ class Hypergraph:
         return self.edges.get_cell_property(edge, node, prop_name)
 
     def get_properties(self, id, level=None, prop_name=None):
+        # method on each property store and incidence store.
         """Returns an object's specific property or all properties
 
         Parameters
@@ -783,6 +799,7 @@ class Hypergraph:
 
     @warn_nwhy
     def get_linegraph(self, s=1, edges=True):
+        # method on incidence store
         """
[…]

[…]
