### 2024-01-17 pnnl/HyperNetX#135 @szhorvat
[joss] Support forum available?
Does HyperNetX have a support forum or chatroom that is more appropriate for general discussion and usage questions than GitHub issues? Or do you prefer such questions to be asked in the issue tracker?
