### 2024-01-17 pnnl/HyperNetX#137 @szhorvat
[joss] Most operations fail on empty hypergraph
**Describe the bug**

Most operations fail on empty hypergraph.

**To Reproduce**

```
hh = hnx.Hypergraph()
hn.draw(hh) # errors
hh.is_connected() # errors
hh.bipartite() # errors
```

Almost all functions / methods I tried threw a non-intuitive error.

HNX 2.1.3

Ref https://github.com/openjournals/joss-reviews/issues/6016
