### 2024-01-17 pnnl/HyperNetX#138 @szhorvat
[joss] Can HNX represent isolated nodes?
Until #137 is answered, I'll just ask questions here. I hope that's okay.

Can HNX represent isolated nodes? If yes, how can I construct a hypergraph with isolated nodes? If not, can you point to the section of the docs that explains this? I'm sure I'm missing something.

I noticed that when removing edges using `.remove_edges()` (docs missing), nodes with no remaining incident edges also disappear.

Ref: https://github.com/openjournals/joss-reviews/issues/6016
