### 2024-01-17 pnnl/HyperNetX#139 @szhorvat
[joss] Documentation contains unrendered LaTeX math
Some documentation pages contain `$`-delimited LaTeX notation that is not rendered.

Example:

https://hypernetx.readthedocs.io/en/latest/modularity.html

Ref: https://github.com/openjournals/joss-reviews/issues/6016
