### 2024-01-17 pnnl/HyperNetX#140 @szhorvat
[joss] Documentation example does not complete after > 5 min runtime
First of all, I must say that it is very nice that the documentation of several functions comes with examples.

However, one of the first examples I tried did not complete after more than a 5 minute runtime (M2 MacBook Pro):

https://hypernetx.readthedocs.io/en/latest/algorithms/algorithms.html#algorithms.contagion.contagion_animation

When an example takes a very long time to run, this should be noted in the docs.

Ref: https://github.com/openjournals/joss-reviews/issues/6016
