### 2024-01-17 pnnl/HyperNetX#138 @bonicim
The current version of HNX does not support representing isolated nodes. This may be a feature in a future version.
