### 2024-01-18 pnnl/HyperNetX#141 @szhorvat
[joss] Random hypergraph generators need more detailed documentation
The [random hypergraph generators](https://hypernetx.readthedocs.io/en/latest/algorithms/algorithms.html#module-algorithms.generative_models) are very sparsely documented at the moment. The documentation only contains links to papers (and the name of the implementor) but it does not explain what these functions actually do. I feel that in most of these cases it would be possible to provide a precise explanation of the generation procedure within a _short_ paragraph.
