### 2024-01-18 pnnl/HyperNetX#138 @szhorvat
Is this mentioned in the documentation or tutorials? If not, it would be nice to clarify it.
