### 2024-01-18 #6016 @szhorvat
@brendapraggastis When citing igraph in the paper, could you please spell it with a lowercase "i" and also include https://arxiv.org/abs/2311.10260 (which will soon become the main citation point, replacing the old DOI-less paper)?
