### 2024-01-18 pnnl/HyperNetX#142 @szhorvat
[joss] More clarity on (optional) dependencies
I followed the installation instructions in https://hypernetx.readthedocs.io/en/latest/install.html and installed into a new virtual environment from PyPI using `pip install hypernetx`. When I load the package, the following message is shown:

```
>>> import hypernetx
 No module named 'igraph'. If you need to use hypernetx.algorithms.hypergraph_modularity, please install additional packages by running the following command: pip install .['all']
```

This is not an ideal user experience, and is likely confusing to some new users. They may be wondering: "Did I install correctly? Will some essential functionality not be working?"

I suggest:

 - Include more information in the documentation about what dependencies are requires for significant submodules
 - Since most people install from PyPI, include instructions on how install all optional dependencies with a single command
 - The warning message on package load gives instructions which only work when installing from sources. This is not user friendly. Most people would install from PyPI
 - Would it make sense to issue this warning only when people actually try to use `hypernetx.algorithms.hypergraph_modularity` instead of on package load? This is a more common practice.

Ref: https://github.com/openjournals/joss-reviews/issues/6016
