### 2024-01-18 8045e3f
updating descriptions on hypergraph and entityset methods
added which new classes will be used (e.g., incidencestore or propertystore)
