### 2024-01-19 #6016 @szhorvat
@brendapraggastis Let's set the end of the week (Sunday) as the time point for finalizing the review. I'm sorry that this took so long. I am still playing with the library and will open issues until then.

Could you please answer here?  https://github.com/pnnl/HyperNetX/issues/135  This is not something for you to fix. I just need to know the best channels to ask for help.

You should treat all issues I open as suggestions for now.

Something that I do believe should be fixed in the paper (not library) is a better summary of related and alternative software. At the moment you are simply referencing a different paper that is: (1) over two years old in a rapidly developing area (2) is not actually a review paper on the topic (3) is paywalled and I cannot access it (can you email it to me please?)  I would expect not just a list of alternatives, but some sort of useful guidance for users on what packages to look at given their application area, and how HyperNetX differs from these. I understand that JOSS uses a very constraining word limit, but you are not at the word limit, and if the paper is to be useful to readers, this is an important area to improve. Looking at how other related software handle this, XGI has a list of related/alternative packages right in their README: https://github.com/xgi-org/xgi
