### 2024-01-19 pnnl/HyperNetX#143 @szhorvat
[joss] HyperNetX Widget not working
The HyperNetX Widget is prominently mentioned in both the documentation and the JOSS paper. https://hypernetx.readthedocs.io/en/latest/widget.html Unfortunately I wasn't able to get it working. Can you help with this?

The paper states that the widget works in Jupyter notebooks. After evaluating the first cell from [the demo notebook](https://github.com/pnnl/hypernetx-widget/blob/master/notebooks/example%20widget.ipynb), I see the following output only:

![image](https://github.com/pnnl/HyperNetX/assets/1212871/99dfe804-0cc2-4b56-9a00-324184cd9235)

No interactive widget appears.

I am using:

 - Jupyter Lab 4.0.8
 - HNX 2.1.3 and hnxwdget 0.1.1b3 (both from PyPI)
 - Python 3.11

The documentation goes into great detail on how to operate the graphical interface provided by the widget, but it is silent on how to make the widget appear and what code to run to create the widget (i.e. the prerequisites to enabling the described GUI). This should be remedied. https://hypernetx.readthedocs.io/en/latest/widget.html

Ref: https://github.com/openjournals/joss-reviews/issues/6016
