### 2024-01-19 #6016 @bonicim
@szhorvat I have answered https://github.com/pnnl/HyperNetX/issues/135. You can now start a general discussion [here](https://github.com/pnnl/HyperNetX/discussions).
