### 2024-01-19 #6016 @brendapraggastis
> Before I can check off this item:
>
> > * Contribution and authorship: Has the submitting author (@brendapraggastis) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> I wanted to note that only 3 out of the 8 authors seem to have made commits to the repository, https://github.com/pnnl/HyperNetX/graphs/contributors (unless I failed to match up some GitHub user names). Authorship policy is up to the journal (and personally I certainly do not think that authorship should require code contributions). Therefore I leave it up to the editors whether they need further clarifications on this.

All listed authors have contributed to the repository. Ongoing code adjustments and documentation must be reviewed by our company before public release so the true commit history isn't visible on Github.
