### 2024-01-19 pnnl/HyperNetX#144 @szhorvat
[joss] Contributor guide should be more specific, and shouldn't suggest Rust as the language
The current contribution guide is quite non-specific and doesn't explain the specificities of contributing to _this_ package. All the links go to generic sources, such as the GitHub documentation. Is there anything particular about contributing to this project that would be useful to mention?

It seems to suggest using Rust as the programming language. Is this a mistake (perhaps a leftover from the original source of this guide, that was missed during editing)?

When I submitted a very minor PR to this project in July (https://github.com/pnnl/HyperNetX/pull/117), it was mentioned that I should have targeted the `develop` branch and that this information will be included in the contributor guide. It's still not mentioned in the current version though.

https://github.com/pnnl/HyperNetX/blob/0ed183145155f90dd225da176c047b0b3e1cc1ac/CONTRIBUTING.md

Ref: https://github.com/openjournals/joss-reviews/issues/6016
