### 2024-01-19 pnnl/HyperNetX#145 @szhorvat
[joss] Testing instructions are incomplete in the documentation
The testing guide contains a complete list of instructions in the README, but not in the documentation, here:

https://hypernetx.readthedocs.io/en/latest/install.html#running-tests

Specifically, `pip install .'[testing]` is missing.

I don't think it is necessary to include testing instructions in the usage documentation (as opposed to development guide), but if you do include it, please make sure that it is complete. This is quite useful e.g. for package maintainers.

Ref: https://github.com/openjournals/joss-reviews/issues/6016
