### 2024-01-19 pnnl/HyperNetX#146 @szhorvat
[joss] Deprecation warning from NetworkX when calling diameter functions
**Describe the bug**

Calling `.*diameter()` functions triggers a deprecation warning from NetworkX

**To Reproduce**

```
H = hnx.Hypergraph([{1,2,3},{2,3}])
H.node_diameters()
```

```
/Users/szhorvat/miniconda3/envs/hnx/lib/python3.11/site-packages/hypernetx/classes/hypergraph.py:1857: DeprecationWarning:

The scipy.sparse array containers will be used instead of matrices
in Networkx 3.0. Use `from_scipy_sparse_array` instead.
  G = nx.from_scipy_sparse_matrix(A)
```

Note that NetworkX 3.0 has been released over a year ago (this virtualenv has 2.8.8)

**Environment (please complete the following information):**
 - HNX 2.1.3
 - NetworkX 2.8.8

Ref: https://github.com/openjournals/joss-reviews/issues/6016
