### 2024-01-22 pnnl/HyperNetX#143 @bonicim
I reproduced the issue using the same configuration. I have tried to use the widget on both Jupyter notebook and Jupyter lab. We are currently working on resolving this issue.
