### 2024-01-23 pnnl/HyperNetX#143 @bonicim
@szhorvat After working with the team on this issue, it appears that `hnxwidget` will now only work on the Anaconda platform (i.e use a conda environment instead of a vanilla `venv` virtual environment). You are right that the documentation needs to be updated and I will start work on that. In the meanwhile, below are my notes on how to use the widget. Also attached is a screenshot of what to expect on Jupyter notebook.

----------Notes---------
# How to run HNX widget

`hnxwidget` is currently in beta and will only work on Jupyter Notebook 6.5.x. It is currently not yet supported on Jupyter Lab. Moreover, `hnxwidget` is currently supported only on the [Anaconda platform](https://www.anaconda.com/); therefore, it is highly recommended to use a `conda` environment to use this widget. Do not use python's builtin `venv` module or `virtualenv`.

## Prerequisites

* conda 23.11.x
* python 3.11.x
* jupyter notebook 6.5.4
* ipywidgets 7.6.5

## Instructions
Open a new shell and run the following commands.

```
# update conda
conda update conda

# activate the base environment
conda activate

# install hypernetx and hnxwidget
pip install hypernetx hnxwidget

# install jupyter notebook and extensions
conda install -y -c anaconda notebook
conda install -y -c conda-forge jupyter_contrib_nbextensions

# install and enable the hnxwidget on jupyter
jupyter nbextension install --py --symlink --sys-prefix hnxwidget
jupyter nbextension enable --py --sys-prefix hnxwidget

# install ipykernel and use it to add the base environment to jupyter notebook
conda install -y -c anaconda ipykernel
python -m ipykernel install --user --name=base

# start the notebook
jupyter-notebook
```

Refs:
[How to add your Conda environment to your jupyter notebook in just 4 steps \| by Noopur R K | Medium](https://medium.com/@nrk25693/how-to-add-your-conda-environment-to-your-jupyter-notebook-in-just-4-steps-abeab8b8d084)

![Screenshot 2024-01-23 at 15 11 28](https://github.com/pnnl/HyperNetX/assets/6901848/b23c7b9b-5b95-481e-a8a1-7cd9278f83bd)
