### 2024-01-24 #6016 @brendapraggastis
@danielskatz and @szhorvat - it looks like the only thing pending is a better description of other hypergraph libraries. This will take some time depending on how in depth a comparison we make.
@szhorvat suggests "I would expect not just a list of alternatives, but some sort of useful guidance for users on what packages to look at given their application area, and how HyperNetX differs from these."
I want to make sure we fairly represent the other libraries and a deep dive into their inner workings will take time. I will discuss this with my team and follow up with you later this week.
