### 2024-01-24 83a3723
update incidence view and hypview
