### 2024-01-26 #6016 @danielskatz
@brendapraggastis - thanks for your update. However, in addition to what you say, I also see a 11 open issues that are tagged with this review, I think all from @szhorvat. It's unclear to me how many of these are blockers for @szhorvat (vs suggestions for the future), but given the unchecked review items in [@szhorvat's checklist](https://github.com/openjournals/joss-reviews/issues/6016#issuecomment-1792342485), I assume some of them do need to be addressed. @szhorvat - perhaps you can make it clear which of these issues fall into which category?
