### 2024-01-28 #6016 @szhorvat
I would like to suggest fixing the following before publication:

 - https://github.com/pnnl/HyperNetX/issues/145 (quick and easy doc fix)
 - https://github.com/pnnl/HyperNetX/issues/144 I consider the contributor guide important for any open source project.
 - https://github.com/pnnl/HyperNetX/issues/143 : This can't be fixed quickly, and it wouldn't be reasonable to ask for this to be fixed before publishing. But given that this feature is prominently mentioned in the paper, and currently it is quite difficult to get it working, could the authors please comment on what the plans are going forward?  Will it be supported and made easy to use, or will it be kept in its current state or dropped?
 - https://github.com/pnnl/HyperNetX/issues/142 This is not a major issue, but again it is a quick fix, and I'd like to see it corrected for the sake of user friendliness.

> I want to make sure we fairly represent the other libraries and a deep dive into their inner workings will take time.

To be clear, I am not asking for a deep dive, after all this is not a review paper. However, currently there is no guidance at all for readers about what alternatives exist and what makes HNX different. There is a figure (Fig 2) with no main text reference, and a citation to Kurte et al, which I still wasn't able to access.

With that out of the way, I wanted to say that HNX looks like a great package that is breaking new ground. Modelling systems using higher order networks is still a relatively new, and quickly developing approach. I enjoyed exploring HNX and I'm looking forward to seeing significant development in the coming years.
