### 2024-01-28 #6016 @IvanIsCoding
@szhorvat https://rdcu.be/dw897 I was able to share the Kurte et al. paper. I hope that helps
