### 2024-02-01 fac0670
Pull request #137: Update installation instructions for widget
Merge in HYP/hypernetx from HYP-384-hnx-widget-not-rendering-on-jupyter-lab to develop

* commit '53dc408129d1fc2ea8bff3b9620589fe179c510c':
  Update installation instructions for widget
