### 2024-02-01 73f9c2e
Added the contain_hyper_edges flag to extend hulls (rubber bands) around hyperedge positions. Useful for when the bipartite graph is passed in, e.g. with_additional_edges=H.bipartite()
