### 2024-02-02 7e5a4a6
HYP-392 Update post-installation steps documentation
