### 2024-02-02 pnnl/HyperNetX#144 @bonicim
Agreed. Will revise the contributor guide.
