### 2024-02-02 40bc072
HYP-391 Update Contributing guide
