### 2024-02-02 0f9c227
HYP-389 Fix optional dependencies; update documentation
