+++ setup.cfg
@@ -93,7 +93,7 @@ testing =
 	igraph>=0.10.4
 tutorials =
 	jupyter>=1.0
-	python-igraph>=0.10.4
+	igraph>=0.10.4
 	partition-igraph>=0.0.6
 	celluloid>=0.2.0
 widget =
