### 2024-02-02 4ebf35b
HYP-389 Remove testing dependency; update test
