+++ setup.cfg
@@ -80,7 +80,6 @@ linting =
 testing =
 	pytest>=7.2.2
 	pytest-cov>=4.1.0
-	pytest-lazy-fixture>=0.6.3
 	pytest-xdist>=3.2.1
 	pytest-env
 	tox>=4.4.11
