+++ tox.ini
@@ -17,7 +17,6 @@ deps =
 	celluloid>=0.2.0
 	igraph>=0.10.4
 	nbmake>=1.4.1
-    pytest-lazy-fixture>=0.6.3
     pytest-xdist>=3.2.1
     partition-igraph>=0.0.6
 allowlist_externals = env
