### 2024-02-02 pnnl/HyperNetX#143 @bonicim
The HyperNetX widget is currently in beta with limitations on the Jupyter environment in which it may be used.  It is being actively worked on. Look for improvements and an expanded list of usable environments in a future release.

In addition, updates to the documentation are currently in progress and are planned to be published in the next week or so.
