### 2024-02-03 #6016 @bonicim
> * [[joss] Testing instructions are incomplete in the documentation pnnl/HyperNetX#145](https://github.com/pnnl/HyperNetX/issues/145) (quick and easy doc fix)

Will fix.

> * [[joss] Contributor guide should be more specific, and probably shouldn't suggest Rust as the language pnnl/HyperNetX#144](https://github.com/pnnl/HyperNetX/issues/144) I consider the contributor guide important for any open source project.

Will fix.

> * [[joss] HyperNetX Widget not working pnnl/HyperNetX#143](https://github.com/pnnl/HyperNetX/issues/143) : This can't be fixed quickly, and it wouldn't be reasonable to ask for this to be fixed before publishing. But given that this feature is prominently mentioned in the paper, and currently it is quite difficult to get it working, could the authors please comment on what the plans are going forward?  Will it be supported and made easy to use, or will it be kept in its current state or dropped?

The HyperNetX widget is currently in beta with limitations on the Jupyter environment in which it may be used.  It is being actively worked on. Look for improvements and an expanded list of usable environments in a future release.

In addition, updates to the documentation are currently in progress and are planned to be published in the next week or so.

> * [[joss] More clarity on (optional) dependencies pnnl/HyperNetX#142](https://github.com/pnnl/HyperNetX/issues/142) This is not a major issue, but again it is a quick fix, and I'd like to see it corrected for the sake of user friendliness.

Will fix.
