### 2024-02-06 9d89948
Update property_store, tests; add factory stubs
