+++ hypernetx/classes/tests/test_propertystore.py
@@ -1,16 +1,0 @@
-from hypernetx.classes.property_store import (
-    PropertyStore,
-    PropertyStoreFactory,
-    DataFramePropertyStore,
-    DictPropertyStore,
-)
-
-
-def test_property_store_factory(harry_potter):
-    level = 0
-    harry_potter.set_random_wandweights()
-    data = harry_potter.dataframe
-
-    ps = PropertyStoreFactory.from_dataframe(level, data)
-
-    assert isinstance(ps, DataFramePropertyStore)
