+++ hypernetx/classes/tests/test_property_store.py
@@ -1,0 +1,4 @@
+from hypernetx.classes.property_store import PropertyStore
+
+
+def test_property_store(): ...
