### 2024-02-12 3c09b18
Coding incidence store and adding incidence store tests
