+++ hypernetx/classes/incidence_store.py
@@ -1,152 +1,160 @@
+import pandas as pd
 
-from copy import copy,deepcopy
-
-class IncidenceStore(ABC):
+class IncidenceStore:
     """
     Incidence store object that stores and accesses (multi) incidences with standard methods.
 
     Parameters
     ----------
-    data : Iterable mapping of ordered pairs (edge,node) to data properties
-    aggregate_by : method for handling properties associated with duplicate pairs,
-        may be given as a dictionary keyed on properties
-        
-    Methods
-    --------
-    __iter__
-    __len__
-    __contains__
-    __getitem__
-    incidence_matrix
-    collapseitem
-    restrict_to
-    dimensions
-    membership
-    elements
-    dual
+    data : Two column pandas dataframe of nodes and edges, respectively.
     """
     
-    def __init__(self, data, edge_level, node_level):
+    def __init__(self, data):
         """
-        Multi index for (Incidence) Property Store with additional
-        functionality
+        Initiate data in self as the two column pandas dataframe provided through factory method.
 
         Parameters
         ----------
         data : _type_
             collection of ordered pairs
         """
-        self._data = incidence_data  # Data with incidence pairs and attributes
-
-        ### class does not allow duplicate pairs so either the first pair
-        # and its data will be used or the aggregate_by method will be
-        # used. these should be documented for behavior 
+        # initiate self with data (pandas dataframe) with duplicate incidence pairs removed.
+        self._data = data.drop_duplicates()  
+        
         pass
-
     
     def __iter__(self):
         """
         Iterator over the incidence pairs of the hypergraph
-
-        """
         
-        return iter(self._data)
+        Returns
+        -------
+        iter of tuples
+            Iterator over incidence pairs (tuples) in the hypergraph.
+        """
+        #itertuples provides iterator over rows in a dataframe with index as false to not return index 
+        #and name as None to return a standard tuple.
+        return iter(self._data.itertuples(index=False, name=None))
 
     def __len__(self):
         """
         Total number of incidences
-
+        
         Returns
         -------
         int
-
+            Number of incidence pairs in the hypergraph.
         """
-        
-        # 
+
         return len(self._data)
     
     def __contains__(self, incidence_pair):
         """
-        First, check if the incidence pair is of length two.        
-        Then, checks if it exists in incidence pairs.
+        Checks if an incidence pair exists in the incidence pairs dataframe.
+        First, this checks if the incidence pair is of length two.        
+        Then, it checks if it exists in the incidence pairs.
+        
+        Parameters
+        ----------
+        incidence_pair : tuple
+           Incidence pair that is a tuple (or array-like object; e.g., list or array) of length two.
 
+        Returns
+        -------
+        bool
+            True if incidence pair exists in incidence store.
         """
-                
-        pass
+        df = self._data
+        
+        #verify the incidence pair is of length two. Otherwise, pair does not exist.
+        if len(incidence_pair) == 2:
+            node, edge = incidence_pair[0], incidence_pair[1]
+            # check if first element in pair (node) exists in 'nodes' column anywhere
+            # and check if second element of pair (edge) exists in 'edges' column anywhere.
+            does_contain = ((df['nodes'] == node) & (df['edges'] == edge)).any()
+            return does_contain
+        else:
+            return False
     
-    # #### Do we need both of these??? Or can we have a single getitem??
-    # def __getitem__(self, incidence_key):
-    #     """
-    #     key is incidence key (e.g., index) and returns incidence pair and attributes
-
-    #     """
-    #     ### let getitem retrieve single line or multiple lines depending on if 
-    #     ### incidence_key is a hashable key for the instance, or an ordered pair of an
-    #     ### incidence. Should replace get_incidence_attributes
-    #     return self._data[incidence_key]
[…]
