### 2024-02-12 dbc189f
Update widget instructions; add screnshot
