### 2024-02-12 62f51a1
updating tests to function correctly.
All incidence store tests passing
