+++ hypernetx/classes/factory.py
@@ -1,62 +1,157 @@
-from typing import Union, Any
+from typing import Any, Tuple, Sequence
 from hypernetx.classes.property_store import PropertyStore
+from hypernetx.classes.incidence_store import IncidenceStore
 
 import pandas as pd
 
+# HNX
+# ability to lookup properties
+
+# return a loc/ series
+# set a attribute into dictionary
+# key, value pair
+
+# setattr affects the class attributes
+# get item, set item (
+# id, key, value
+
+# remove getattr, setattr
+# setattr(id, key, val)
+# use MultiIndex instead of tuple for Id
+# index based on edgeid, nodeid, multiindex (no dupes)
+
 
 def create_property_store_df(
-    data: pd.DataFrame, level: Union[int, str]
-) -> PropertyStore:
-    """Parse dataframe and create PropertyStore instances
+    data: pd.DataFrame, properties: pd.DataFrame, data_cols: Sequence[str | int]
+) -> Tuple[PropertyStore, PropertyStore, PropertyStore, IncidenceStore]:
+    """Parse data and properties and create PropertyStore instances
 
     Returns:
 
-    PropertyStore: id's map to properties
     """
-    ...
 
+    # create three dataframes (no multiindex) - nodes, edge, incidence pairs,
+    # 4 helper methods
+    # create df for nodes
+    # create df for edge
+    # create df for incidence pairs
+    # create df for incidence store
 
-def create_property_store_dict(
-    data: dict[Union[int, str], dict[Any, Any]]
-) -> PropertyStore:
-    """Parse dictionary and create PropertyStore instance
+    # By default: create properties even if not provided
+    # for laterenable properties flag; set default true
 
-    Returns:
+    # step 1: create initial dfp
 
-    PropertyStore: id's map to properties
-    """
-    ...
+    # step 2:
+    # TODO: Fix to check the shape of properties or redo properties format
+    column_map = {
+        old: new
+        for old, new in zip(
+            (level_col, id_col, misc_col),
+            (*self.properties.index.names, self._misc_props_col),
+        )
+        if old is not None
+    }
+    props = props.rename(columns=column_map)
+    props = props.rename_axis(index=column_map)
 
+    # Step 3:
+    # create the properties from dataframe
 
-def create_property_store_incidence_pairs_df(data: pd.DataFrame) -> PropertyStore:
-    """Parse dataframe and create PropertyStore instance for incidence pairs
+    # names of property table idx-levels for level and item id, respectively
+    # ``item`` used instead of ``id`` to avoid redefining python built-in func `id`
+    level, item = self.properties.index.names
+    if props.index.nlevels > 1:  # props has MultiIndex
+        # drop all idx-levels from props other than level and id (if present)
+        extra_levels = [
+            idx_lev for idx_lev in props.index.names if idx_lev not in (level, item)
+        ]
+        props = props.reset_index(level=extra_levels)
 
-    Returns:
+    try:
+        # if props index is already in the correct format,
+        # enforce the correct idx-level ordering
+        props.index = props.index.reorder_levels((level, item))
+    except AttributeError:  # props is not in (level, id) MultiIndex format
+        # if the index matches level or id, drop index to column
+        if props.index.name in (level, item):
+            props = props.reset_index()
+        index_cols = [item]
+        if level in props:
+            index_cols.insert(0, level)
+        try:
+            props = props.set_index(index_cols, verify_integrity=True)
+        except ValueError:
+            warnings.warn(
+                "duplicate (level, ID) rows will be dropped after first occurrence"
+            )
+            props = props.drop_duplicates(index_cols)
+            props = props.set_index(index_cols)
 
[…]
