+++ _sources/hypergraph101.rst.txt
@@ -1,0 +1,465 @@
+.. _hypergraph101:
+
+===============================================
+A Gentle Introduction to Hypergraph Mathematics
+===============================================
+
+
+Here we gently introduce some of the basic concepts in hypergraph
+modeling. We note that in order to maintain this “gentleness”, we will
+be mostly avoiding the very important and legitimate issues in the
+proper mathematical foundations of hypergraphs and closely related
+structures, which can be very complicated. Rather we will be focusing on
+only the most common cases used in most real modeling, and call a graph
+or hypergraph **gentle** when they are loopless, simple, finite,
+connected, and lacking empty hyperedges, isolated vertices, labels,
+weights, or attributes. Additionally, the deep connections between
+hypergraphs and other critical mathematical objects like partial orders,
+finite topologies, and topological complexes will also be treated
+elsewhere. When it comes up, below we will sometimes refer to the added
+complexities which would attend if we weren’t being so “gentle”. In
+general the reader is referred to [1,2] for a less gentle and more
+comprehensive treatment.
+
+Graphs and Hypergraphs
+======================
+
+Network science is based on the concept of a **graph**
+:math:`G=\langle V,E\rangle` as a system of connections between
+entities. :math:`V` is a (typically finite) set of elements, nodes, or
+objects, which we formally call **“vertices”**, and :math:`E` is a set
+of pairs of vertices. Given that, then for two vertices
+:math:`u,v \in V`, an **edge** is a set :math:`e=\{u,v\}` in :math:`E`,
+indicating that there is a connection between :math:`u` and :math:`v`.
+It is then common to represent :math:`G` as either a Boolean **adjacency
+matrix** :math:`A_{n \times n}` where :math:`n=|V|`, where an
+:math:`i,j` entry in :math:`A` is 1 if :math:`v_i,v_j` are connected in
+:math:`G`; or as an **incidence matrix** :math:`I_{n \times m}`, where
+now also :math:`m=|E|`, and an :math:`i,j` entry in :math:`I` is now 1
+if the vertex :math:`v_i` is in edge :math:`e_j`.
+
+.. _f1:
+.. figure:: images/exgraph.png
+    :class: with-border
+    :width: 300
+    :align: center
+
+    An example graph, where the numbers are edge IDs.
+
+.. _t1:
+.. list-table:: Adjacency matrix :math:`A` of a graph.
+   :header-rows: 1
+   :align: center
+
+   * -
+     - Andrews
+     - Bailey
+     - Carter
+     - Davis
+   * - Andrews
+     - 0
+     - 1
+     - 1
+     - 1
+   * - Bailey
+     - 1
+     - 0
+     - 1
+     - 0
+   * - Carter
+     - 1
+     - 1
+     - 0
+     - 1
+   * - Davis
+     - 1
+     - 0
+     - 1
+     - 1
+
+.. _t2:
+.. list-table:: Incidence matrix :math:`I` of a graph.
+   :header-rows: 1
+   :align: center
+
+   * -
+     - 1
+     - 2
+     - 3
+     - 4
+     - 5
+   * - Andrews
+     - 1
+     - 1
+     - 0
+     - 1
+     - 0
+   * - Bailey
+     - 0
+     - 0
+     - 0
+     - 1
+     - 1
+   * - Carter
+     - 0
+     - 1
+     - 1
+     - 0
+     - 1
+   * - Davis
+     - 1
+     - 0
+     - 1
+     - 0
+     - 0
+
+
+.. _label3:
+.. figure:: images/biblio_hg.png
+    :class: with-border
+    :width: 400
+    :align: center
+
[…]
