### 2024-02-15 #6016 @bonicim
@danielskatz Updates pushed to master branch and reflected in latest release, 2.1.4. PyPi has also been updated.

Additional comments on other hypergraph libraries are forthcoming.
