### 2024-02-15 8f8f40d
slowly adding methods using proxies for incidence store and property store.
