+++ hypernetx/classes/hyp_view.py
@@ -20,10 +20,52 @@ from hypernetx.classes import EntitySet
 from hypernetx.exception import HyperNetXError
 from hypernetx.utils.decorators import warn_nwhy
 from hypernetx.classes.helpers import merge_nested_dicts, dict_depth
+from hypernetx.classes.incidence_store import IncidenceStore as IS
+
 
 __all__ = ["Hypergraph"]
 T = TypeVar("T", bound=Union[str, int])
 
+
+##################### PROXY CLASSES FOR CONSTRUCTION
+class PropertyStore(object):
+    def __init__(self,dfp=None):
+        if dfp is not None:
+            self._dataframe = dfp
+            self._dataframe['properties'] = [{} for idx in self.dataframe.index]
+        else:
+            self._dataframe = pd.DataFrame(columns=['id','weight','properties'])
+    def __call_(self):
+        return self
+
+    @property
+    def dataframe(self):
+        return self._dataframe
+
+class IncidenceStore():
+    def __init__(self,df):
+        self._dataframe = df
+        self.proxy = IS(df)
+            
+    def __call__(self):
+        return self
+    
+    def __iter__(self):
+        return self.proxy.__iter__()
+    
+    @property
+    def edges(self):
+        return self.proxy.edges
+    
+    @property
+    def nodes(self):
+        return self.proxy.nodes
+
+    @property
+    def dataframe(self):
+        return self._dataframe
+    
+######################################################
 
 class HypergraphView(object):
     """
@@ -31,7 +73,7 @@ class HypergraphView(object):
     meta data for hypergraph. Provides methods matching EntitySet 
     methods in previous versions.
     """
-    def __init__(incidence_store,level,property_store=None):
+    def __init__(self,incidence_store,level,property_store=None):
         """
         _summary_
 
@@ -44,176 +86,175 @@ class HypergraphView(object):
         property_store : _type_, optional
             _description_, by default None
         """
-
-        self._incidences = incidence_store
+        self.name = "test"
+        self._incidence_store = incidence_store
         ### incidence store needs index or columns
         self._level = level  ## edges, nodes, or incidence pairs
-        if property_store is not None:
-            self._properties = property_store
-            ### property store needs index column = keys
-        else:
-            self._properties = hnx.PropertyStore()  
-            ### if no properties and level 0 or 1, 
-            ### create property store that 
-            ### returns weight 1 on every call for a weight 
-            ### and empty properties otherwise.
+        self._properties = property_store or PropertyStore()
 
+        # self._properties = PropertyStore()  
+        ### if no properties and level 0 or 1, 
+        ### create property store that 
+        ### returns weight 1 on every call for a weight 
+        ### and empty properties otherwise.
 
+    @property
+    def dataframe(self):
+        return self._incidence_store.dataframe
+    
     def __iter__(self):
         """
         Defined by level store
+        iterate over the associated level in the incidence store
         """
-        pass
+        level = self._level
+        if level == 0:
+            return self._incidence_store.edges.__iter__()
+        elif level == 1:
+            return self._incidence_store.nodes.__iter__()
+        elif level == 2:
+            return self._incidence_store.__iter__()
 
     def __len__(self):
         """
-        Defined by level store
+        Defined by incidence store 
         """
-        pass
+        level = self._level
+        if 
 
-    def __contains__(self,item):
-        """
-        Defined by level store
+    # def __contains__(self,item):
+    #     """
+    #     Defined by level store
 
-        Parameters
-        ----------
-        item : _type_
-            _description_
-        """
-        pass
+    #     Parameters
+    #     ----------
+    #     item : _type_
+    #         _description_
+    #     """
+    #     pass
 
     
-    def to_dataframe(self):
-        """
[…]
