### 2024-02-15 #6016 @bonicim
I will look into the pdf generation error. I have an idea of what happened; will rerun the command once I've updated the `paper` branch.
