### 2024-02-16 9a2e0f1
Implement DataFramePropertyStore; add tests
