+++ hypernetx/classes/property_store.py
@@ -1,14 +1,14 @@
-from dataclasses import dataclass
 from abc import ABC, abstractmethod
-from typing import Union, Any
+from typing import Any, Optional
 
 import pandas as pd
 
+UID = "uid"
+WEIGHT = "weight"
+PROPERTIES = "properties"
 
-@dataclass
+
 class PropertyStore(ABC):
-    data: Any = None
-
     @property
     @abstractmethod
     def properties(self) -> Any:
@@ -16,12 +16,13 @@ class PropertyStore(ABC):
         ...
 
     @abstractmethod
-    def get_properties(self, key) -> dict[Any, Any]:
+    def get_properties(self, uid) -> dict[Any, Any]:
         """Get all properties of an item
 
         Parameters
         ----------
-        key : str | int
+        uid : tuple[str | int] | tuple[str | int, str | int ]
+            edge name, node name, or edge-node pair
 
         Returns
         -------
@@ -32,7 +33,7 @@ class PropertyStore(ABC):
         Raises
         ------
         KeyError
-            if (`key`) is not in :attr:`properties`,
+            if (`uid`) is not in :attr:`properties`,
 
         See Also
         --------
@@ -41,13 +42,13 @@ class PropertyStore(ABC):
         ...
 
     @abstractmethod
-    def get_property(self, key, prop_name) -> Any:
+    def get_property(self, uid, prop_name) -> Any:
         """Get a property of an item
 
         Parameters
         ----------
-        key : str | int
-            name of an item
+        uid : tuple[str | int] | tuple[str | int, str | int ]
+            edge name, node name, or edge-node pair
         prop_name : str | int
             name of the property to get
 
@@ -62,7 +63,7 @@ class PropertyStore(ABC):
         Raises
         ------
         KeyError
-            if (`key`) is not in :attr:`properties`,
+            if (`uid`) is not in :attr:`properties`,
 
         See Also
         --------
@@ -71,13 +72,13 @@ class PropertyStore(ABC):
         ...
 
     @abstractmethod
-    def set_property(self, key, prop_name, prop_val) -> None:
-        """Set a property of an item
+    def set_property(self, uid, prop_name, prop_val) -> None:
+        """Set a property of an item in the 'properties' dictionary
 
         Parameters
         ----------
-        key : str | int
-            name of an item
+        uid : tuple[str | int] | tuple[str | int, str | int ]
+            edge name, node name, or edge-node pair
         prop_name : str | int
             name of the property to set
         prop_val : any
@@ -85,8 +86,9 @@ class PropertyStore(ABC):
 
         Raises
         ------
-        ValueError
-            If `key` is not in :attr:`properties`
+        KeyError
+            If (`uid`) is not in :attr:`properties`
+
 
         See Also
         --------
@@ -112,12 +114,15 @@ class PropertyStore(ABC):
         """
         ...
 
-    def __getitem__(self, key: Union[str, int]) -> dict:
-        """Returns the common attributes (e.g. weight) and properties of a key in the underlying data table
+    def __getitem__(self, uid) -> dict:
+        """Returns the common attributes (e.g. weight) and properties of an item in the underlying data table
+
+        Returns same result as get_property
 
         Parameters
         ----------
-        key : str | int
+        uid : tuple[str | int] | tuple[str | int, str | int ]
+            edge name, node name, or edge-node pair
 
         Returns
         -------
@@ -125,12 +130,13 @@ class PropertyStore(ABC):
         """
         ...
 
-    def __contains__(self, key: Union[str, int]) -> bool:
-        """Returns true if key is in the underlying data table; false otherwise
+    def __contains__(self, uid) -> bool:
[…]
