### 2024-02-16 pnnl/HyperNetX#142 @szhorvat
Is `pip install hypenetx['all']` something that should be suggested?
