### 2024-02-16 pnnl/HyperNetX#143 @szhorvat
Can you please clarify what you mean here?

> It is highly recommended to use the base environment provided by Anaconda because Anaconda’s package management system, conda, will resolve dependencies when HyperNetX and HyperNetXWidget are installed.

Is it suggesting to install HNX and HNXWidget in the base environment instead of a dedicated environment? If so, why? My impression was that installing everything in the base environment was not considered good practice.
