### 2024-02-19 c66e952
finished draft dataframe factory method.
See example at bottom of factory.py for use example.
