+++ hypernetx/classes/factory.py
@@ -1,157 +1,221 @@
-from typing import Any, Tuple, Sequence
-from hypernetx.classes.property_store import PropertyStore
-from hypernetx.classes.incidence_store import IncidenceStore
 
 import pandas as pd
+import networkx as nx
+import numpy as np
 
-# HNX
-# ability to lookup properties
+'''
+Thoughts:
+    Why do we not just follow networkx as an example and ditch all of the factory methods 
+    and only be able to add nodes and edges and incidences in a specific format? We would still have everything but it would be empty to start.
+'''
 
-# return a loc/ series
-# set a attribute into dictionary
-# key, value pair
+# In[ ]:
+#-------------------------------------------------------------------------------------------------
+# Individual factory methods for incidence stores
+#-------------------------------------------------------------------------------------------------
+'''notes: 
+   * worried that choosing edge names as integers is a problem if they match node names. This is the case for iterable of iterables. Assuming this is fine.
+   * for IoI or any method without attributes do I return none or trivial property stores? Could in the future plan for 
+     "simple" hypergraph that does not have properties. For now I am returning trivial property store.s (i.e., no properties except weight = 1)
+     have default method that returns weight or properties if there are no properties provided. 
+     Should still instantiate an empty dataframe with the correct columns though.
+   * should the attributes on the repeated instances also be combined? Not sure exactly how this would be done. I'm leaning towards just doing weight for now and then
+     maybe adding a lossless additional combined attributes later on if it is needed.'
+     
+   * can have a properties input into the hypergraph creation for incidences, edges, and nodes. 
+   Also, incidence properties can be retrieved directly from the set system, but only for the incidences.
+'''
 
-# setattr affects the class attributes
-# get item, set item (
-# id, key, value
+def to_incidence_store_from_iterable_of_iterables(iterable_of_iterables):
+    edge_list = []# edge component of incidence pairs as list
+    node_list = []# node component of incidence pairs as list
+    for edge, nodes_of_edge in enumerate(iterable_of_iterables):
+        for node in nodes_of_edge:
+            #append incidence pair of (edge, node) to edge and node lists, respectively.
+            edge_list += [edge]
+            node_list += [node]
+    incidence_store_dict = {'edges': edge_list, 
+                            'nodes': node_list}
+    #Incidence Store (IS)
+    IS = pd.DataFrame(incidence_store_dict)
+    return IS
 
-# remove getattr, setattr
-# setattr(id, key, val)
-# use MultiIndex instead of tuple for Id
-# index based on edgeid, nodeid, multiindex (no dupes)
-
-
-def create_property_store_df(
-    data: pd.DataFrame, properties: pd.DataFrame, data_cols: Sequence[str | int]
-) -> Tuple[PropertyStore, PropertyStore, PropertyStore, IncidenceStore]:
-    """Parse data and properties and create PropertyStore instances
-
-    Returns:
-
-    """
-
-    # create three dataframes (no multiindex) - nodes, edge, incidence pairs,
-    # 4 helper methods
-    # create df for nodes
-    # create df for edge
-    # create df for incidence pairs
-    # create df for incidence store
-
-    # By default: create properties even if not provided
-    # for laterenable properties flag; set default true
-
-    # step 1: create initial dfp
-
-    # step 2:
-    # TODO: Fix to check the shape of properties or redo properties format
-    column_map = {
-        old: new
-        for old, new in zip(
-            (level_col, id_col, misc_col),
-            (*self.properties.index.names, self._misc_props_col),
-        )
-        if old is not None
-    }
-    props = props.rename(columns=column_map)
[…]
