### 2024-02-21 ec6636f
updated paper with some references to hypergraph libraries.
