### 2024-02-22 #6016 @bonicim
Added section on comparison with other hypergraph libraries.
