### 2024-02-23 c9a94c6
Use MultiIndexDataframe for PropertyStore; update tests; add notebook
