+++ hypernetx/classes/tests/test_property_store.py
@@ -3,81 +3,85 @@ import pandas as pd
 from hypernetx.classes.property_store import PropertyStore, DataFramePropertyStore
 
 
-@pytest.fixture
-def edges() -> list[tuple[str]]:
-    return [("P",), ("R",), ("S",), ("L",), ("M",), ("N",)]
+EDGE_LEVEL = 'e'
+NODE_LEVEL = 'n'
+LEVEL = 'level'
+ID = 'id'
+UUID = 'uuid'
+WEIGHT = 'weight'
+PROPERTIES = 'properties'
+PROPERTIES_COLUMNS = [WEIGHT, PROPERTIES]
+STRENGTH = 'strength'
+HAIR_COLOR = 'hair_color'
+INCIDENCES_PROPERTIES_COLUMNS = [WEIGHT, PROPERTIES, STRENGTH, HAIR_COLOR]
 
 
 @pytest.fixture
-def df_edges(edges) -> pd.DataFrame:
-    uid = edges
-    data = {
-        "uid": uid,
-        "weight": [1.0 for _ in range(len(uid))],
-        "properties": [{} for _ in range(len(uid))],
-    }
-    return pd.DataFrame.from_dict(data)
+def edges() -> list[tuple[str | int, str | int]]:
+    edges = ['I', 'L', 'O', 'P', 'R', 'S']
+    return [(EDGE_LEVEL, e) for e in edges]
 
 
 @pytest.fixture
-def dfps_edges(df_edges) -> PropertyStore:
-    return DataFramePropertyStore(df_edges)
+def nodes():
+    nodes = ['A', 'C', 'E', 'K', 'T1', 'T2', 'V']
+    return [(NODE_LEVEL, n) for n in nodes]
 
 
 @pytest.fixture
-def incident_pairs() -> list[tuple[str, str]]:
-    return [
-        ("P", "K"),
-        ("P", "A"),
-        ("P", "C"),
-        ("R", "A"),
-        ("R", "E"),
-        ("S", "K"),
-        ("S", "A"),
-        ("S", "V"),
-        ("S", "T2"),
-        ("L", "E"),
-        ("L", "C"),
-        ("M", "T2"),
-        ("M", "T1"),
-        ("N", "K"),
-        ("N", "T2"),
-    ]
+def incidences():
+    return [('I', 'K'), ('I', 'T2'), ('L', 'C'), ('L', 'E'), ('O', 'T1'), ('O', 'T2'), ('P', 'A'), ('P', 'C'),
+                  ('P', 'K'), ('R', 'A'), ('R', 'E'), ('S', 'A'), ('S', 'K'), ('S', 'T2'), ('S', 'V')]
 
 
 @pytest.fixture
-def df_incident_pairs(incident_pairs) -> pd.DataFrame:
-    uid = incident_pairs
-    data = {
-        "uid": uid,
-        "weight": [1.0 for _ in range(len(uid))],
-        "properties": [{} for _ in range(len(uid))],
-        "strength": [42 for _ in range(len(uid))],
-        "hair_color": ["red" for _ in range(len(uid))],
-    }
-    return pd.DataFrame.from_dict(data)
+def edges_df(edges) -> pd.DataFrame:
+    index = pd.MultiIndex.from_tuples(edges, names=[LEVEL, ID])
+    data = [(1, {}) for _ in range(len(index))]
+    return pd.DataFrame(data=data, index=index, columns=PROPERTIES_COLUMNS)
 
 
 @pytest.fixture
-def dfps_incident_pairs(df_incident_pairs) -> PropertyStore:
-    return DataFramePropertyStore(df_incident_pairs)
+def nodes_df(nodes) -> pd.DataFrame:
[…]
