+++ hypernetx/classes/property_store.py
@@ -1,9 +1,10 @@
 from abc import ABC, abstractmethod
-from typing import Any, Optional
+from typing import Any
+from collections.abc import Hashable
 
 import pandas as pd
 
-UID = "uid"
+UUID = "uuid"
 WEIGHT = "weight"
 PROPERTIES = "properties"
 
@@ -12,23 +13,29 @@ class PropertyStore(ABC):
     @property
     @abstractmethod
     def properties(self) -> Any:
-        """Properties assigned to items in the underlying data table"""
+        """Properties assigned to all items in the underlying data table
+
+        Returns
+        -------
+        Object containing all properties of the underlying data table
+        """
         ...
 
     @abstractmethod
-    def get_properties(self, uid) -> dict[Any, Any]:
+    def get_properties(self, uid: Hashable) -> dict[Any, Any]:
         """Get all properties of an item
 
         Parameters
         ----------
-        uid : tuple[str | int] | tuple[str | int, str | int ]
-            edge name, node name, or edge-node pair
+        uid : Hashable
+            uid is the index used to fetch all its properties
 
         Returns
         -------
         prop_vals : dict
             ``{named property: property value, ...,
-            misc. property column name: {property name: property value}}``
+            properties: {property name: property value}}``
+
 
         Raises
         ------
@@ -42,13 +49,13 @@ class PropertyStore(ABC):
         ...
 
     @abstractmethod
-    def get_property(self, uid, prop_name) -> Any:
+    def get_property(self, uid: Hashable, prop_name: str | int) -> Any:
         """Get a property of an item
 
         Parameters
         ----------
-        uid : tuple[str | int] | tuple[str | int, str | int ]
-            edge name, node name, or edge-node pair
+        uid : Hashable
+            uid is the index used to fetch its property
         prop_name : str | int
             name of the property to get
 
@@ -72,13 +79,13 @@ class PropertyStore(ABC):
         ...
 
     @abstractmethod
-    def set_property(self, uid, prop_name, prop_val) -> None:
-        """Set a property of an item in the 'properties' dictionary
+    def set_property(self, uid: Hashable, prop_name: str | int, prop_val: Any) -> None:
+        """Set a property of an item in the 'properties' colelction
 
         Parameters
         ----------
-        uid : tuple[str | int] | tuple[str | int, str | int ]
-            edge name, node name, or edge-node pair
+        uid : Hashable
+            uid is the index used to set its property
         prop_name : str | int
             name of the property to set
         prop_val : any
@@ -95,7 +102,18 @@ class PropertyStore(ABC):
         get_property, get_properties
         """
 
-    def __iter__(self) -> iter:
+    def __iter__(self):
+        """
+        Returns an iterator object for iterating over data in PropertyStore
+
+        Returns:
+        -------
+        DataIterator:
+            An iterator object
+        """
+        return self
+
+    def __iter__(self):
         """Returns an iterator over items in the underlying data table
 
         Returns
@@ -121,8 +139,8 @@ class PropertyStore(ABC):
 
         Parameters
         ----------
-        uid : tuple[str | int] | tuple[str | int, str | int ]
-            edge name, node name, or edge-node pair
+        Hashable
+            uid is the index used to set its property
 
         Returns
         -------
@@ -135,8 +153,8 @@ class PropertyStore(ABC):
 
         Parameters
         ----------
-        uid : tuple[str | int] | tuple[str | int, str | int ]
-            edge name, node name, or edge-node pair
+        Hashable
+            uid is the index used to set its property
 
         Returns
         -------
@@ -146,96 +164,164 @@ class PropertyStore(ABC):
 
 
 class DataFramePropertyStore(PropertyStore):
[…]
