+++ hypernetx/classes/tests/test_entityset_on_dict.py
@@ -1,56 +1,52 @@
 import numpy as np
 import pytest
 
-from pytest_lazyfixture import lazy_fixture
-
 from hypernetx.classes import EntitySet
 
 
 @pytest.mark.parametrize(
     "entity, data, data_cols, labels",
     [
-        (lazy_fixture("sbs_dict"), None, (0, 1), None),
+        (("sbs_dict"), None, (0, 1), None),
         (
-            lazy_fixture("sbs_dict"),
+            ("sbs_dict"),
             None,
             (0, 1),
-            lazy_fixture("sbs_labels"),
+            ("sbs_labels"),
         ),  # labels are ignored if entity is provided
-        (lazy_fixture("sbs_dict"), None, ["edges", "nodes"], None),
-        (lazy_fixture("sbs_dict"), lazy_fixture("sbs_data"), (0, 1), None),
-        (None, lazy_fixture("sbs_data"), (0, 1), lazy_fixture("sbs_labels")),
+        ("sbs_dict", None, ["edges", "nodes"], None)
     ],
 )
 class TestEntitySBSDict:
     """Tests on different use cases for combination of the following params: entity, data, data_cols, labels"""
 
-    def test_size(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
+    def test_size(self, entity, data, data_cols, labels, sbs, request):
+        es = EntitySet(entity=request.getfixturevalue(entity), data=data, data_cols=data_cols, labels=labels)
         assert es.size() == len(sbs.edgedict)
 
     # check all the EntitySet properties
-    def test_isstatic(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
+    def test_isstatic(self, entity, data, data_cols, labels, sbs, request):
+        es = EntitySet(entity=request.getfixturevalue(entity), data=data, data_cols=data_cols, labels=labels)
         assert es.isstatic
 
-    def test_uid(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
+    def test_uid(self, entity, data, data_cols, labels, sbs, request):
+        es = EntitySet(entity=request.getfixturevalue(entity), data=data, data_cols=data_cols, labels=labels)
         assert es.uid is None
 
-    def test_empty(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
+    def test_empty(self, entity, data, data_cols, labels, sbs, request):
+        es = EntitySet(entity=request.getfixturevalue(entity), data=data, data_cols=data_cols, labels=labels)
         assert not es.empty
 
-    def test_uidset(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
+    def test_uidset(self, entity, data, data_cols, labels, sbs, request):
+        es = EntitySet(entity=request.getfixturevalue(entity), data=data, data_cols=data_cols, labels=labels)
         assert es.uidset == {"I", "R", "S", "P", "O", "L"}
 
-    def test_dimsize(self, entity, data, data_cols, labels, sbs):
-        es = EntitySet(entity=entity, data=data, data_cols=data_cols, labels=labels)
+    def test_dimsize(self, entity, data, data_cols, labels, sbs, request):
+        es = EntitySet(entity=request.getfixturevalue(entity), data=data, data_cols=data_cols, labels=labels)
         assert es.dimsize == 2
 
-    def test_elements(self, entity, data, data_cols, labels, sbs):
[…]
