### 2024-02-23 pnnl/HyperNetX#142 @bonicim
`pip install hypenetx['all']` installs Jupyter, allowing the user to open one of the tutorials in the repo; it also installs the dependencies needed for the contagion and modularity sub-modules. The [documentation](https://hypernetx.readthedocs.io/en/latest/install.html#installing-from-source) recommends this.
