### 2024-02-23 pnnl/HyperNetX#142 @szhorvat
Could you please clarify where the documentation mentions this command? I can only see `pip install -e .['all']` in the section about installing from source.

My impression was that the target audience of this library is not just developers, but mainly researchers. I expect most of them won't want to install from source and won't necessarily notice that `pip install hypenetx['all']` is also a valid command.
