### 2024-02-23 pnnl/HyperNetX#142 @bonicim
I see your point about the target audience being researchers. I will update the installing from PyPi instructions  to include a snippet about `pip install hypernetx['all']`
