### 2024-02-26 15954cd
updates incorporating new property and incidence stores
