+++ hypernetx/classes/hyp_view.py
@@ -25,80 +25,80 @@ from hypernetx.classes.incidence_store import IncidenceStore as IS
 __all__ = ["HypergraphView"]
 
 ##################### PROXY CLASSES FOR CONSTRUCTION
-class PropertyStore(object):
-    """
-    Wrapper for a pandas dataframe. Minimal logic but limits changes
-    to 
+# class PropertyStore(object):
+#     """
+#     Wrapper for a pandas dataframe. Minimal logic but limits changes
+#     to 
 
-    Parameters
-    ----------
-    object : _type_
-        _description_
-    """
-    def __init__(self,dfp=None, level=0):
-        if dfp is not None:
-            self._dataframe = dfp
-            if 'properties' not in dfp.columns:
-                self._dataframe['properties'] = [{} for idx in self.dataframe.index]
-            else:
-                dfp.properties.fillna({})
-            if 'weight' not in dfp.columns:
-                self._dataframe.weight = 1
-            else:
-                dfp.weight.fillna(1)
-            if level in [0,1]:
-                self._dataframe = self._dataframe.set_index(self._dataframe.columns[0])
-            elif level in [2]:
-                self._dataframe = self._dataframe.set_index([self._dataframe.columns[0],self._dataframe.columns[1]])
-        else:
-            self._dataframe = pd.DataFrame(columns=['weight','properties'])
+#     Parameters
+#     ----------
+#     object : _type_
+#         _description_
+#     """
+#     def __init__(self,dfp=None, level=0):
+#         if dfp is not None:
+#             self._dataframe = dfp
+#             if 'properties' not in dfp.columns:
+#                 self._dataframe['properties'] = [{} for idx in self.dataframe.index]
+#             else:
+#                 dfp.properties.fillna({})
+#             if 'weight' not in dfp.columns:
+#                 self._dataframe.weight = 1
+#             else:
+#                 dfp.weight.fillna(1)
+#             if level in [0,1]:
+#                 self._dataframe = self._dataframe.set_index(self._dataframe.columns[0])
+#             elif level in [2]:
+#                 self._dataframe = self._dataframe.set_index([self._dataframe.columns[0],self._dataframe.columns[1]])
+#         else:
+#             self._dataframe = pd.DataFrame(columns=['weight','properties'])
     
     
-    def __call__(self):
-        return self
+#     def __call__(self):
+#         return self
     
-    def __iter__(self):
-        return iter(self._dataframe.index)
+#     def __iter__(self):
+#         return iter(self._dataframe.index)
     
-    def __len__(self):
-        return len(self._dataframe)
+#     def __len__(self):
+#         return len(self._dataframe)
 
-    @property
-    def dataframe(self):
-        return self._dataframe
+#     @property
+#     def dataframe(self):
+#         return self._dataframe
     
-    def get_property(self, uid, prop_name):
-        prop_val = None
-        df = self.dataframe
-        try:
-            prop_val = df.loc[uid][prop_name]
-        except KeyError:
-            prop_val = df.loc[uid]['properties'].get(prop_name,None)
-        return prop_val
+#     def get_property(self, uid, prop_name):
+#         prop_val = None
+#         df = self.dataframe
+#         try:
+#             prop_val = df.loc[uid][prop_name]
+#         except KeyError:
+#             prop_val = df.loc[uid]['properties'].get(prop_name,None)
+#         return prop_val
 
-class IncidenceStore(IS):
-    def __init__(self,df):
-        super().__init__(df)
-        # self._dataframe = self._data
-        # self.proxy = IS(df)
+# class IncidenceStore(IS):
+#     def __init__(self,df):
+#         super().__init__(df)
+#         # self._dataframe = self._data
[…]
