### 2024-02-26 fd73f5b
small modifications to integrate better with existing code and hyp_view
