### 2024-02-26 4857324
small changes for better integration with hyp_view and existing code
