+++ hypernetx/classes/property_store.py
@@ -10,161 +10,163 @@ PROPERTIES = "properties"
 
 
 class PropertyStore(ABC):
-    @property
-    @abstractmethod
-    def properties(self) -> Any:
-        """Properties assigned to all items in the underlying data table
+    def __init__(self):
+        pass
+    # @property
+    # @abstractmethod
+    # def properties(self) -> Any:
+    #     """Properties assigned to all items in the underlying data table
 
-        Returns
-        -------
-        Object containing all properties of the underlying data table
-        """
-        ...
+    #     Returns
+    #     -------
+    #     Object containing all properties of the underlying data table
+    #     """
+    #     ...
 
-    @abstractmethod
-    def get_properties(self, uid: Hashable) -> dict[Any, Any]:
-        """Get all properties of an item
+    # @abstractmethod
+    # def get_properties(self, uid: Hashable) -> dict[Any, Any]:
+    #     """Get all properties of an item
 
-        Parameters
-        ----------
-        uid : Hashable
-            uid is the index used to fetch all its properties
+    #     Parameters
+    #     ----------
+    #     uid : Hashable
+    #         uid is the index used to fetch all its properties
 
-        Returns
-        -------
-        prop_vals : dict
-            ``{named property: property value, ...,
-            properties: {property name: property value}}``
-
-
-        Raises
-        ------
-        KeyError
-            if (`uid`) is not in :attr:`properties`,
-
-        See Also
-        --------
-        get_property, set_property
-        """
-        ...
-
-    @abstractmethod
-    def get_property(self, uid: Hashable, prop_name: str | int) -> Any:
-        """Get a property of an item
-
-        Parameters
-        ----------
-        uid : Hashable
-            uid is the index used to fetch its property
-        prop_name : str | int
-            name of the property to get
-
-        Returns
-        -------
-        prop_val : any
-            value of the property
-
-        None
-            if property not found
-
-        Raises
-        ------
-        KeyError
-            if (`uid`) is not in :attr:`properties`,
-
-        See Also
-        --------
-        get_properties, set_property
-        """
-        ...
-
-    @abstractmethod
-    def set_property(self, uid: Hashable, prop_name: str | int, prop_val: Any) -> None:
-        """Set a property of an item in the 'properties' colelction
-
-        Parameters
-        ----------
-        uid : Hashable
-            uid is the index used to set its property
-        prop_name : str | int
-            name of the property to set
-        prop_val : any
-            value of the property to set
-
-        Raises
-        ------
-        KeyError
-            If (`uid`) is not in :attr:`properties`
+    #     Returns
+    #     -------
+    #     prop_vals : dict
+    #         ``{named property: property value, ...,
+    #         properties: {property name: property value}}``
 
 
-        See Also
-        --------
-        get_property, get_properties
-        """
+    #     Raises
+    #     ------
+    #     KeyError
+    #         if (`uid`) is not in :attr:`properties`,
 
-    def __iter__(self):
-        """
-        Returns an iterator object for iterating over data in PropertyStore
+    #     See Also
+    #     --------
+    #     get_property, set_property
+    #     """
+    #     ...
 
-        Returns:
-        -------
-        DataIterator:
-            An iterator object
-        """
-        return self
+    # @abstractmethod
+    # def get_property(self, uid: Hashable, prop_name: str | int) -> Any:
+    #     """Get a property of an item
 
-    def __iter__(self):
-        """Returns an iterator over items in the underlying data table
+    #     Parameters
+    #     ----------
+    #     uid : Hashable
+    #         uid is the index used to fetch its property
[…]
