### 2024-02-26 97ff5bd
added all optional args for factory method
