+++ hypernetx/classes/factory.py
@@ -26,20 +26,6 @@ Thoughts:
    Also, incidence properties can be retrieved directly from the set system, but only for the incidences.
 '''
 
-def to_incidence_store_from_iterable_of_iterables(iterable_of_iterables):
-    edge_list = []# edge component of incidence pairs as list
-    node_list = []# node component of incidence pairs as list
-    for edge, nodes_of_edge in enumerate(iterable_of_iterables):
-        for node in nodes_of_edge:
-            #append incidence pair of (edge, node) to edge and node lists, respectively.
-            edge_list += [edge]
-            node_list += [node]
-    incidence_store_dict = {'edges': edge_list, 
-                            'nodes': node_list}
-    #Incidence Store (IS)
-    IS = pd.DataFrame(incidence_store_dict)
-    return IS
-
 def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_col = 'edges', node_col = 'nodes'):
     #change column names to edges and nodes.
     column_renaming_dict = {edge_col: 'edges', node_col: 'nodes'}
@@ -53,11 +39,20 @@ def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_c
 # Individual factory methods for property stores
 #-------------------------------------------------------------------------------------------------
 def to_property_store_from_dataframe(properties, property_type, 
-                                     edge_col = 'edges', node_col = 'nodes', weight_col = 'weight', misc_properties_col = 'properties',
-                                     default_cell_weight = 1.0, default_edge_weight = 1.0, default_node_weight = 1.0):
+                                     edge_col = 'edges', node_col = 'nodes', 
+                                     misc_cell_properties = 'properties', misc_properties = 'properties',
+                                     weight_prop = None, edge_weight_prop = 'weight', node_weight_prop = 'weight',
+                                     cell_weight_col = 'weight',
+                                     cell_weights = 1.0, default_edge_weight = 1.0, default_node_weight = 1.0):
+    
+    
+    
     properties_df_dict = {}
     
-    #set UID column
+    
+    
+    
+    #create a multi-index with level and id columns.
     if property_type == 'cell_properties':
         incidence_pairs = np.array(properties[[edge_col, node_col]]) #array of incidence pairs to use as UIDs.
         indices = [tuple(incidence_pair) for incidence_pair in incidence_pairs]
@@ -67,48 +62,134 @@ def to_property_store_from_dataframe(properties, property_type,
     elif property_type == 'node_properties': 
         node_names = np.array(properties[[node_col]]) #array of edges to use as UIDs.
         indices = [tuple(['n', node_name[0]]) for node_name in node_names]
-        
     multi_index = pd.MultiIndex.from_tuples(indices, names=['level', 'id'])
     
-    #set property columns including weight and misc props.
     
+    
+    
+
     #get names of property columns provided that are not the edge or node columns
     property_columns = set(list(properties.columns)) - set([edge_col, node_col])
     for prop in property_columns:
         properties_df_dict[prop] = properties.loc[:, prop]
         
-    #set weight property column if not already provided
-    default_weights = {'cell_properties': default_cell_weight, 'edge_properties': default_edge_weight, 'node_properties': default_node_weight}
-    if weight_col not in property_columns:
-        properties_df_dict[weight_col] = [default_weights[property_type]]*len(indices)
-    
-    print(misc_properties_col)
-    #set misc properties column if not already provided
-    if misc_properties_col not in property_columns:
-        properties_df_dict[misc_properties_col] = [{}]*len(indices)
         
-    #change not of misc props column to "properties"
-    if misc_properties_col != 'properties':
-        # Pop the value associated with properties key and store it
[…]
