### 2024-02-26 b73e21f
Update property_store tests
