### 2024-02-27 #6016 @brendapraggastis
@szhorvat and @danielskatz the latest version had references to the three most contemporary libraries to HNX that we know of along with their references. Hopefully this version is acceptable for publication.
