### 2024-02-28 d90b2c3
Add hypergraph_view factory method
