+++ hypernetx/classes/hyp_view.py
@@ -20,6 +20,7 @@ from hypernetx.exception import HyperNetXError
 from hypernetx.utils.decorators import warn_nwhy
 from hypernetx.classes.helpers import merge_nested_dicts, dict_depth
 from hypernetx.classes.incidence_store import IncidenceStore as IS
+from hypernetx.classes.factory import restructure_data
 
 
 __all__ = ["HypergraphView"]
@@ -28,7 +29,7 @@ __all__ = ["HypergraphView"]
 # class PropertyStore(object):
 #     """
 #     Wrapper for a pandas dataframe. Minimal logic but limits changes
-#     to 
+#     to
 
 #     Parameters
 #     ----------
@@ -52,21 +53,21 @@ __all__ = ["HypergraphView"]
 #                 self._dataframe = self._dataframe.set_index([self._dataframe.columns[0],self._dataframe.columns[1]])
 #         else:
 #             self._dataframe = pd.DataFrame(columns=['weight','properties'])
-    
-    
+
+
 #     def __call__(self):
 #         return self
-    
+
 #     def __iter__(self):
 #         return iter(self._dataframe.index)
-    
+
 #     def __len__(self):
 #         return len(self._dataframe)
 
 #     @property
 #     def dataframe(self):
 #         return self._dataframe
-    
+
 #     def get_property(self, uid, prop_name):
 #         prop_val = None
 #         df = self.dataframe
@@ -81,17 +82,17 @@ __all__ = ["HypergraphView"]
 #         super().__init__(df)
 #         # self._dataframe = self._data
 #         # self.proxy = IS(df)
-            
+
 #     # def __call__(self):
 #     #     return self
-    
+
 #     # def __iter__(self):
 #     #     return self.proxy.__iter__()
-    
+
 #     # @property
 #     # def edges(self):
 #     #     return self.proxy.edges
-    
+
 #     # @property
 #     # def nodes(self):
 #     #     return self.proxy.nodes
@@ -99,13 +100,13 @@ __all__ = ["HypergraphView"]
 #     @property
 #     def dataframe(self):
 #         return self._data
-    
+
 ######################################################
 
 class HypergraphView(object):
     """
-    Wrapper for Property and Incidence Stores holding structural and 
-    meta data for hypergraph. Provides methods matching EntitySet 
+    Wrapper for Property and Incidence Stores holding structural and
+    meta data for hypergraph. Provides methods matching EntitySet
     methods in previous versions. Only nodes and edges in the Incidence
     Store will be seeable in this view.
     """
@@ -129,7 +130,7 @@ class HypergraphView(object):
 
         else:
             self._props = PropertyStore()
-        
+
 
         ### incidence store needs index or columns
         if level == 0 :
@@ -139,28 +140,28 @@ class HypergraphView(object):
         elif level == 2 :
             self._items = self._store.data.values
 
-        # self._properties = PropertyStore()  
-        ### if no properties and level 0 or 1, 
-        ### create property store that 
-        ### returns weight 1 on every call for a weight 
+        # self._properties = PropertyStore()
+        ### if no properties and level 0 or 1,
+        ### create property store that
+        ### returns weight 1 on every call for a weight
         ### and empty properties otherwise.
 
     @property
     def level(self):
         return self._level
-    
+
     @property
     def incidence_store(self):
         return self._store
-    
+
     @property
     def property_store(self):
         return self._props
-        
+
     @property
     def properties(self):
         return self._props.properties
-    
+
 
     # @property
     # def levelset(self):
@@ -179,8 +180,8 @@ class HypergraphView(object):
     #         return self._store.nodes.unique()
     #     elif level == 2:
     #         return self._store.dataframe
-        
-    
+
+
[…]
