### 2024-02-28 #6016 @szhorvat
I was still waiting for a response on a minor comment here: https://github.com/pnnl/HyperNetX/issues/143#issuecomment-1949265449

Perhaps other reviewers can comment on whether what is suggested (i.e. install everything in the conda base environment) is considered good practice. Python technical details are not my area of expertise.

My guiding principle with these comments was to try to make sure that academic users, many of whom might not be very technical, will be well-served by the documentation.
