+++ hypernetx/classes/factory.py
@@ -45,11 +45,72 @@ def to_property_store_from_dataframe(properties, property_type,
                                      cell_weight_col = 'weight',
                                      cell_weights = 1.0, default_edge_weight = 1.0, default_node_weight = 1.0):
     
+    """
+    This function creates a pandas dataframe in the correct format given a 
+    pandas dataframe of either cell, node, or edge properties.
     
+    Parameters
+    ----------
+    
+    properties : dataframe
+        dataframe of properties for either incicences, edges, or nodes
+    
+    property_type : str
+        type of property dataframe provided. 
+        Either 'edge_properties', 'node_properties', or 'cell_properties'
+
+    edge_col : (optional) str | int, default = 0
+        column index (or name) in pandas.dataframe or numpy.ndarray,
+        used for (hyper)edge ids. Will be used to reference edgeids for
+        all set systems.
+
+    node_col : (optional) str | int, default = 1
+        column index (or name) in pandas.dataframe or numpy.ndarray,
+        used for node ids. Will be used to reference nodeids for all set systems.
+
+    cell_weight_col : (optional) str | int, default = None
+        column index (or name) in pandas.dataframe or numpy.ndarray used for
+        referencing cell weights. For a dict of dicts references key in cell
+        property dicts.
+
+    cell_weights : (optional) Sequence[float,int] | int |  float , default = 1.0
+        User specified cell_weights or default cell weight.
+        Sequential values are only used if setsystem is a
+        dataframe or ndarray in which case the sequence must
+        have the same length and order as these objects.
+        Sequential values are ignored for dataframes if cell_weight_col is already
+        a column in the data frame.
+        If cell_weights is assigned a single value
+        then it will be used as default for missing values or when no cell_weight_col
+        is given.
+
+    misc_cell_properties : (optional) str | int, default = None
+        Column name of dataframe corresponding to a column of variable
+        length property dictionaries for the cell. Ignored for other setsystem
+        types.
+
+    misc_properties : (optional) int | str, default = None
+        Column of property dataframes with dtype=dict. Intended for variable
+        length property dictionaries for the objects.
+
+    edge_weight_prop : (optional) str, default = None,
+        Name of property in edge_properties to use for weight.
+
+    node_weight_prop : (optional) str, default = None,
+        Name of property in node_properties to use for weight.
+
+    weight_prop : (optional) str, default = None
+        Name of property in properties to use for 'weight'
+
+    default_edge_weight : (optional) int | float, default = 1
+        Used when edge weight property is missing or undefined.
+
+    default_node_weight : (optional) int | float, default = 1
+        Used when node weight property is missing or undefined
+        
+    """
     
     properties_df_dict = {}
-    
-    
     
     
     #create a multi-index with level and id columns.
@@ -74,6 +135,22 @@ def to_property_store_from_dataframe(properties, property_type,
         properties_df_dict[prop] = properties.loc[:, prop]
         
         
+    #get column names if integer was provided instead
+    if isinstance(edge_col, int):
+        edge_col = properties.columns[edge_col]
+    if isinstance(edge_weight_prop, int):
+        edge_weight_prop = properties.columns[edge_weight_prop]
+    if isinstance(node_col, int):
+        node_col = properties.columns[node_col]
+    if isinstance(node_weight_prop, int):
+        node_weight_prop = properties.columns[node_weight_prop]
+    if isinstance(misc_properties, int):
+        misc_properties = properties.columns[misc_properties]  
+    if isinstance(cell_weight_col, int):
+        cell_weight_col = properties.columns[cell_weight_col] 
+    if isinstance(misc_properties, int):
+        misc_cell_properties = properties.columns[misc_cell_properties] 
+    
         
[…]
