### 2024-02-28 5cdb85c
Allow DataFramePropertyStore to accept regular or multiindex dfs; change properties to misc_properties
