+++ hypernetx/classes/tests/test_property_store.py
@@ -1,15 +1,11 @@
 import pytest
 import pandas as pd
-from hypernetx.classes.property_store import PropertyStore, DataFramePropertyStore
+from hypernetx.classes.property_store import PropertyStore, DataFramePropertyStore, WEIGHT, PROPERTIES
 
 
-EDGE_LEVEL = 'e'
-NODE_LEVEL = 'n'
 LEVEL = 'level'
 ID = 'id'
 UUID = 'uuid'
-WEIGHT = 'weight'
-PROPERTIES = 'properties'
 PROPERTIES_COLUMNS = [WEIGHT, PROPERTIES]
 STRENGTH = 'strength'
 HAIR_COLOR = 'hair_color'
@@ -18,14 +14,12 @@ INCIDENCES_PROPERTIES_COLUMNS = [WEIGHT, PROPERTIES, STRENGTH, HAIR_COLOR]
 
 @pytest.fixture
 def edges() -> list[tuple[str | int, str | int]]:
-    edges = ['I', 'L', 'O', 'P', 'R', 'S']
-    return [(EDGE_LEVEL, e) for e in edges]
+    return ['I', 'L', 'O', 'P', 'R', 'S']
 
 
 @pytest.fixture
 def nodes():
-    nodes = ['A', 'C', 'E', 'K', 'T1', 'T2', 'V']
-    return [(NODE_LEVEL, n) for n in nodes]
+    return ['A', 'C', 'E', 'K', 'T1', 'T2', 'V']
 
 
 @pytest.fixture
@@ -36,15 +30,15 @@ def incidences():
 
 @pytest.fixture
 def edges_df(edges) -> pd.DataFrame:
-    index = pd.MultiIndex.from_tuples(edges, names=[LEVEL, ID])
-    data = [(1, {}) for _ in range(len(index))]
+    data = [(1, {}) for _ in edges]
+    index = pd.Index(edges, name=ID)
     return pd.DataFrame(data=data, index=index, columns=PROPERTIES_COLUMNS)
 
 
 @pytest.fixture
 def nodes_df(nodes) -> pd.DataFrame:
-    index = pd.MultiIndex.from_tuples(nodes, names=[LEVEL, ID])
-    data = [(1, {}) for _ in range(len(index))]
+    data = [(1, {}) for _ in nodes]
+    index = pd.Index(nodes, name=ID)
     return pd.DataFrame(data=data, index=index, columns=PROPERTIES_COLUMNS)
 
 
@@ -79,11 +73,11 @@ def test_properties(edges_dfps, edges):
 @pytest.mark.parametrize(
     "fixture, uid, expected",
     [
-        ("edges_dfps", ("e", "P"), {"weight": 1.0, "properties": dict()}),
+        ("edges_dfps", "P", {"weight": 1.0, PROPERTIES: dict()}),
         (
             "incidences_dfps",
             ("S", "A"),
-            {"weight": 1.0, "properties": dict(), "hair_color": "red", "strength": 42},
+            {"weight": 1.0, PROPERTIES: dict(), "hair_color": "red", "strength": 42},
         ),
     ],
 )
@@ -95,29 +89,27 @@ def test_get_properties(fixture, uid, expected, request):
 
 @pytest.mark.parametrize(
     "fixture, uid",
-    [("edges_dfps", ("e", "NEMO")), ("incidences_dfps", ("NE", "MO"))],
+    [("edges_dfps", "NEMO"), ("incidences_dfps", ("NE", "MO"))],
 )
 def test_get_properties_raises_key_error(fixture, uid, request):
     with pytest.raises(KeyError) as exc_info:
         ps = request.getfixturevalue(fixture)
         ps.get_properties(uid)
 
-    assert (
[…]
