### 2024-02-28 e2858fe
Revert "Add hypergraph_view factory method"
This reverts commit 058b31911a8dd4953a429319a91cf521ea3a1eda.
