### 2024-02-29 pnnl/HyperNetX#143 @drj11
I agree that installing everything in the base environment is not best practice, but it's still a very widely established practice. As long it's documented, i personally would not withhold a publication recommendation on that basis.
