### 2024-03-01 bafb065
Update optional dependencies; cleanup Makefile
