+++ setup.cfg
@@ -92,7 +92,6 @@ testing =
 tutorials =
 	jupyter>=1.0
 	igraph>=0.10.4
-	partition-igraph>=0.0.6
 	celluloid>=0.2.0
 	shutup>=0.2.0
 widget =
@@ -104,20 +103,12 @@ documentation =
 	sphinx-rtd-theme>=1.2.1
 	sphinx-autobuild>=2021.3.14
 	sphinx-copybutton>=0.5.1
+	nb2plots>=0.6.1
 packaging =
 	build>=0.10.0
 	twine>=4.0.2
 	setuptools>=67.6.1
 	tox>=4.4.11
 all =
-	sphinx>=6.2.1
-	nb2plots>=0.6.1
-	sphinx-rtd-theme>=1.2.0
-	sphinx-autobuild>=2021.3.14
-	sphinx-copybutton>=0.5.1
-	pytest>=7.2.2
-	pytest-cov>=4.1.0
-	jupyter>=1.0
-	igraph>=0.10.4
-	partition-igraph>=0.0.6
 	celluloid>=0.2.0
+	igraph>=0.10.4
