+++ Makefile
@@ -25,58 +25,65 @@ flake8:
 format:
 	@$(PYTHON3) -m black hypernetx
 
+
 ## Tests
 
+.PHONY: pre-commit
 pre-commit:
 	pre-commit install
 	pre-commit run --all-files
 
-
+.PHONY: test
 test:
 	coverage run --source=hypernetx -m pytest
 	coverage report -m
 
+.PHONY: test-ci
 test-ci:
 	@$(PYTHON3) -m tox
 
+.PHONY: test-ci-stash
 test-ci-stash: lint-deps lint pre-commit test-deps test-ci
 
+
+.PHONY: test-ci-github
 test-ci-github: lint-deps lint pre-commit ci-github-deps test-deps test-ci
 
-.PHONY: pre-commit test test-ci, test-ci-stash, test-ci-github
 
 ## Continuous Deployment
 ## Assumes that scripts are run on a container or test server VM
-
 ### Publish to PyPi
+
+.PHONY: publish-deps
 publish-deps:
 	@$(PYTHON3) -m pip install -e .'[packaging]' --use-pep517
 
+.PHONY: build-dist
 build-dist: publish-deps clean
 	@$(PYTHON3) -m build --wheel --sdist
 	@$(PYTHON3) -m twine check dist/*
 
 ## Assumes the following environment variables are set: TWINE_USERNAME, TWINE_PASSWORD, TWINE_REPOSITORY_URL,
 ## See https://twine.readthedocs.io/en/stable/#environment-variables
+.PHONY: publish-to-pypi
 publish-to-pypi: publish-deps build-dist
 	@echo "Publishing to PyPi"
 	$(PYTHON3) -m twine upload dist/*
 
-.PHONY: build-dist publish-to-pypi publish-deps
 
 ### Update version
 
+.PHONY: version-deps
 version-deps:
 	@$(PYTHON3) -m pip install .'[releases]' --use-pep517
 
-.PHONY: version-deps
 
 ### Documentation
 
+.PHONY: docs-deps
 docs-deps:
 	@$(PYTHON3) -m pip install .'[documentation]' --use-pep517
 
-.PHONY: docs-deps
 
 ## Tutorials
 
@@ -89,15 +96,17 @@ tutorials:
 	jupyter notebook tutorials
 
 
-
 ## Environment
 
+.PHONY: clean-venv
 clean-venv:
 	rm -rf $(VENV)
 
+.PHONY: clean
 clean:
 	rm -rf .out .pytest_cache .tox *.egg-info dist build
 
+.PHONY: venv
 venv: clean-venv
 	@$(PYTHON3) -m venv $(VENV);
 
@@ -113,10 +122,10 @@ lint-deps:
 format-deps:
 	@$(PYTHON3) -m pip install .'[format]' --use-pep517
 
+.PHONY: test-deps
 test-deps:
 	@$(PYTHON3) -m pip install .'[testing]' --use-pep517
 
+.PHONY: all-deps
 all-deps:
-	@$(PYTHON3) -m pip install -e .'[all]' --use-pep517
-
-.PHONY: clean clean-venv venv all-deps test-deps
+	@$(PYTHON3) -m pip install .'[all]' --use-pep517
