### 2024-03-01 43d35de
Update docs on installation, widget; update Readme and Makefile
