+++ setup.cfg
@@ -103,7 +103,6 @@ documentation =
 	sphinx-rtd-theme>=1.2.1
 	sphinx-autobuild>=2021.3.14
 	sphinx-copybutton>=0.5.1
-	nb2plots>=0.6.1
 packaging =
 	build>=0.10.0
 	twine>=4.0.2
