+++ hypernetx/classes/factory.py
@@ -3,28 +3,10 @@ import pandas as pd
 import networkx as nx
 import numpy as np
 
-'''
-Thoughts:
-    Why do we not just follow networkx as an example and ditch all of the factory methods 
-    and only be able to add nodes and edges and incidences in a specific format? We would still have everything but it would be empty to start.
-'''
-
 # In[ ]:
 #-------------------------------------------------------------------------------------------------
 # Individual factory methods for incidence stores
 #-------------------------------------------------------------------------------------------------
-'''notes: 
-   * worried that choosing edge names as integers is a problem if they match node names. This is the case for iterable of iterables. Assuming this is fine.
-   * for IoI or any method without attributes do I return none or trivial property stores? Could in the future plan for 
-     "simple" hypergraph that does not have properties. For now I am returning trivial property store.s (i.e., no properties except weight = 1)
-     have default method that returns weight or properties if there are no properties provided. 
-     Should still instantiate an empty dataframe with the correct columns though.
-   * should the attributes on the repeated instances also be combined? Not sure exactly how this would be done. I'm leaning towards just doing weight for now and then
-     maybe adding a lossless additional combined attributes later on if it is needed.'
-     
-   * can have a properties input into the hypergraph creation for incidences, edges, and nodes. 
-   Also, incidence properties can be retrieved directly from the set system, but only for the incidences.
-'''
 
 def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_col = 'edges', node_col = 'nodes'):
     #change column names to edges and nodes.
@@ -34,16 +16,21 @@ def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_c
 
 
 
+
+
 # In[ ]:
 #-------------------------------------------------------------------------------------------------
 # Individual factory methods for property stores
 #-------------------------------------------------------------------------------------------------
+
 def to_property_store_from_dataframe(properties, property_type, 
                                      edge_col = 'edges', node_col = 'nodes', 
-                                     misc_cell_properties = 'properties', misc_properties = 'properties',
+                                     misc_cell_properties = 'misc_properties', misc_properties = 'misc_properties',
                                      weight_prop = None, edge_weight_prop = 'weight', node_weight_prop = 'weight',
                                      cell_weight_col = 'weight',
                                      cell_weights = 1.0, default_edge_weight = 1.0, default_node_weight = 1.0):
+    
+    
     
     """
     This function creates a pandas dataframe in the correct format given a 
@@ -117,17 +104,16 @@ def to_property_store_from_dataframe(properties, property_type,
     if property_type == 'cell_properties':
         incidence_pairs = np.array(properties[[edge_col, node_col]]) #array of incidence pairs to use as UIDs.
         indices = [tuple(incidence_pair) for incidence_pair in incidence_pairs]
+        multi_index = pd.MultiIndex.from_tuples(indices, names=['edges', 'nodes'])
     elif property_type == 'edge_properties': 
         edge_names = np.array(properties[[edge_col]]) #array of edges to use as UIDs.
-        indices = [tuple(['e', edge_name[0]]) for edge_name in edge_names]
+        indices = [tuple([edge_name[0]]) for edge_name in edge_names]
+        multi_index = pd.MultiIndex.from_tuples(indices, names=['edges'])
     elif property_type == 'node_properties': 
         node_names = np.array(properties[[node_col]]) #array of edges to use as UIDs.
-        indices = [tuple(['n', node_name[0]]) for node_name in node_names]
-    multi_index = pd.MultiIndex.from_tuples(indices, names=['level', 'id'])
-    
-    
-    
-    
[…]
