### 2024-03-02 pnnl/HyperNetX#142 @bonicim
Have just updated installation steps on the docs: https://hypernetx.readthedocs.io/en/latest/install.html#id1
