### 2024-03-02 pnnl/HyperNetX#143 @bonicim
Have updated the environment setup steps for the widget: https://hypernetx.readthedocs.io/en/latest/widget.html#installation
