### 2024-03-02 #6016 @szhorvat
> @brendapraggastis When citing igraph in the paper, could you please spell it with a lowercase "i" and also include https://arxiv.org/abs/2311.10260 (which will eventually become the main citation point, replacing the old DOI-less paper)?

Could you please correct the uppercase "i" to lowercase before finalizing? I'll write up my report soon.
