### 2024-03-02 #6016 @szhorvat
This is my final report. Please note that I'll be away, and unresponsive, during the following week.

HNX is an ambitious package that is breaking new ground within a rapidly developing field. I am happy to see such packages published in JOSS, and I am looking forward to seeing further HNX development.

Some suggestions to the authors:

**Openness of development** HNX is open-source in the sense that it has an open source license, and that it is technically accepting contributions on GitHub. However, the development process is opaque, and does not project an image that encourages contributions. Looking at [the list of PRs](https://github.com/pnnl/HyperNetX/pulls?q=is%3Apr), the majority of non-trivial ones were closed with a comment that [they were handled internally](https://github.com/pnnl/HyperNetX/pull/122#issuecomment-1804923058). There are very few truly external contributions, and those that exist did not leave a mark on the git commit history (i.e. the name of the actual contributor wasn't recorded). I think HNX could truly benefit from transitioning to more welcoming practices. Consider making development more open and crediting contributors in some way.

**HNX widget** The HNX widget is featured prominently in the paper, but it turned out to function only under very specific circumstances. The authors improved the documentation and stated that the widget is still in beta [and under active development](https://github.com/pnnl/HyperNetX/issues/143#issuecomment-1924901948). However, there [hasn't been any public development (or release) in 3 years](https://github.com/pnnl/hypernetx-widget). Is this another symptom of the behind-closed-doors development practices?  I am hoping to see significant improvements in the near future.

**Documentation** I found the documentation to be a bit terse for a scientific package (see e.g. https://github.com/pnnl/HyperNetX/issues/141) I know that keeping documentation minimal is very common with Python packages—HNX is not different in this respect. But give the target audience (researchers) and the mathematically non-trivial topics the package deals with, I think that putting more effort into documentation would be one of the biggest improvements HNX could make at the moment. Initiatives like pyOpenSci are already pushing for better documentation practices, I suggest taking a look at their guidelines, or just looking at common practices used in the communities of more scientifically oriented languages (e.g. R or Mathematica).

@bonicim @brendapraggastis Thank you for the patience. I hand over to @danielskatz now.
