### 2024-03-02 #6016 @IvanIsCoding
> @IvanIsCoding, @drj11 - any thoughts on this? See the issue in the comment above for the discussion

For the record on the Jupyter Widget debate.

With regards to packaging, ideally users would only `pip install hnxwidget` and just use it. But I myself have experienced issues with Jupyter Widgets as a user so I imagine that on the packaging side it can be challenging as well. A solution using both `pip` and `conda` is not ideal, so my suggestion is to package `hnxwidget` for Conda.

> > @brendapraggastis When citing igraph in the paper, could you please spell it with a lowercase "i" and also include https://arxiv.org/abs/2311.10260 (which will eventually become the main citation point, replacing the old DOI-less paper)?
>
> Could you please correct the uppercase "i" to lowercase before finalizing? I'll write up my report soon.

I think this is one is on the journal, the template capitalizes everyword. [The author's bibtex is what you would expect.](https://github.com/pnnl/HyperNetX/blob/paper/HNX_Paper/paper.bib#L173). I think if they switch to use `{i}graph` in Bibtex it will preserve the capitalization?
