### 2024-03-02 #6016 @szhorvat
> I think if they switch to use `{i}graph` in Bibtex it will preserve the capitalization?

Yes, that works, but I meant to refer to the mention in the main text :-)

Also, to make sure there is no confusion, I do recommend publishing.
