### 2024-03-02 #6016 @IvanIsCoding
> @IvanIsCoding - having now had all the reviews finish, we can go ahead and publish. Let me know if you want to make and changes at this point based on comments. Once you are ready, I'll proofread the paper, and ask you to perform some archiving steps, then I'll finish the process.

I have no objections for the software. For the paper, my only objections are minor igraph citation details. I agree with the other reviewer on the spelling, it is an easy change to make. Another factor that I agree is that igraph has recently posted https://arxiv.org/abs/2311.10260, so I do agree with @szhorvat's comment to cite the pre-print and incorporate the decade of work spent improving the package.
