### 2024-03-03 #6016 @brendapraggastis
@danielskatz @IvanIsCoding @szhorvat Thanks for catching the Igraph error. We will fix that and the references this week and notify you as soon as the final is ready.
