### 2024-03-04 3174272
updated paper refs for igraph
