+++ HNX_Paper/paper.bib
@@ -169,9 +169,17 @@ month = {10}
   publisher={AcM Press New York}
 }
 
+@misc{antonov2023igraph,
+      title={{i}graph enables fast and robust network analysis across programming languages}, 
+      author={Michael Antonov and Gábor Csárdi and Szabolcs Horvát and Kirill Müller and Tamás Nepusz and Daniel Noom and Maëlle Salmon and Vincent Traag and Brooke Foucault Welles and Fabio Zanini},
+      year={2023},
+      eprint={2311.10260},
+      archivePrefix={arXiv},
+      primaryClass={cs.SI}
+}
 
-@article{csardi2006,
-  title={The igraph software package for complex network research},
+@article{csardi2006igraph,
+  title={The {i}graph software package for complex network research},
   author={Csardi, Gabor and Nepusz, Tamas and others},
   journal={InterJournal, complex systems},
   volume={1695},
