### 2024-03-04 #6016 @bonicim
@danielskatz @IvanIsCoding @szhorvat igraph typo is fixed and igraph reference added.
