### 2024-03-04 #6016 @brendapraggastis
> @danielskatz @IvanIsCoding @szhorvat igraph typo is fixed and igraph reference added.

I believe we are good to go for a final proof. The footnote references will be updated by JOSS, correct?
