### 2024-03-05 1a38168
minor changes in paper.md
