+++ HNX_Paper/paper.bib
@@ -68,7 +68,7 @@
 }
 
 @misc{Berge1973Graphs,
-  title={Graphs and {H}ypergraphs,(translated by {E}dward {M}inieka)},
+  title={Graphs and {H}ypergraphs, (translated by {E}dward {M}inieka)},
   doi={10.1016/s0924-6509(09)x7013-3},
   author={Berge, Claude},
   year={1973},
@@ -203,7 +203,7 @@ month = {10}
   number = {85}, 
   pages = {5162}, 
   author = {Nicholas W. Landry and Maxime Lucas and Iacopo Iacopini and Giovanni Petri and Alice Schwarze and Alice Patania and Leo Torres}, 
-  title = {XGI: A Python package for higher-order interaction networks}, journal = {Journal of Open Source Software} 
+  title = {XGI: A {P}ython package for higher-order interaction networks}, journal = {Journal of Open Source Software} 
 }
 
 @inproceedings{Szufel2019,
