### 2024-03-05 559e695
updated hypergraph outline with suggestions for factory methods
