+++ hypernetx/classes/hypergraph_outline.py
@@ -1,1921 +1,0 @@
-# Copyright © 2018 Battelle Memorial Institute
-# All rights reserved.
-from __future__ import annotations
-
-import warnings
-
-warnings.filterwarnings("default", category=DeprecationWarning)
-
-from copy import deepcopy
-from collections import defaultdict
-from collections.abc import Sequence, Iterable
-from typing import Optional, Any, TypeVar, Union, Mapping, Hashable
-
-import networkx as nx
-import numpy as np
-import pandas as pd
-from networkx.algorithms import bipartite
-from scipy.sparse import coo_matrix, csr_matrix
-
-from hypernetx.classes import EntitySet
-from hypernetx.exception import HyperNetXError
-from hypernetx.utils.decorators import warn_nwhy
-from hypernetx.classes.helpers import merge_nested_dicts, dict_depth
-
-__all__ = ["Hypergraph"]
-T = TypeVar("T", bound=Union[str, int])
-
-
-class Hypergraph:
-
-    def __init__(structural_data, properties):
-
-        #### Use a Factory Method to create 4 stores
-            ## df = Incidence Store from structural data 
-            ## edges,nodes,incidences = Property stores tied to ids in df
-            ## Incidences - uses df, dictionary - assigns keys to incidence 
-            ## pairs and/or removes duplicates, 
-            ## identifies keys for edges and nodes
-        
-        ## Construct 3 HypergraphViews to tie these together?
-            ## Incidences - links keys from 3 stores, and incidences with properties
-            ## Edges - reconciles keys from Incidences against a Property Store
-            ## Nodes - reconciles keys from Incidences against a Property Store
-        
-        # decision tree to route data by type OR user specified type
-        # to correct factory method for each of 3 property stores
-        # and 1 incidence store 
-        # 3 calls to the different factory methods (data => df for Property Store)
-        # 1 call to the incidence factory method (setsystem => df for Incidence Store)
-
-        self._E = HypergraphView(incidencestore,propertystore,level=2)
-        self._edges = HypergraphView(incidencestore,propertystore,level=0)
-        self._nodes = HypergraphView(incidencestore,propertystore,level=1)
-
-        self._dataframe = self.dataframe()
-        self._set_default_state()
-        self._dict_.update(locals())
-
-    @property
-    def edges(self):
-        """
-        Object associated with self._edges.
-
-        Returns
-        -------
-        EntitySet
-        """
-        return self._edges
-
-    @property
-    def nodes(self):
-        """
-        Object associated with self._nodes.
-
-        Returns
-        -------
-        EntitySet
-        """
-        return self._nodes
-
-    @property
-    def dataframe(self):
-        """Returns dataframe of incidence pairs and their properties.
-
-        Returns
-        -------
-        pd.DataFrame
-        """
-        return self._E.dataframe
-
-    @property ### TBD
-    def properties(self):
-        """Returns dataframe of edge and node properties.
-
-        Returns
-        -------
-        pd.DataFrame or Dictionary?
-        """
-        ## Concatenate self._edges and self._nodes into a dataframe
-        
-
-    @property
-    def edge_props(self):
-        """Dataframe of edge properties
-        indexed on edge ids
-
-        Returns
-        -------
-        pd.DataFrame or Dictionary?
-        """
-        return self._edges.dataframe
-
-    @property
-    def node_props(self):
-        """Dataframe of node properties
-        indexed on node ids
-
-        Returns
-        -------
-        pd.DataFrame or Dictionary?
-        """
-        return self._nodes.dataframe
-
-    @property
-    def incidence_dict(self):
-        """
-        Dictionary keyed by edge uids with values the uids of nodes in each
-        edge
-
-        Returns
-        -------
-        dict
-
-        """
-        return self._E.to_dict()
-
-    @property
[…]
