### 2024-03-05 8943b7d
added outline of hypergraph constructor in hypergraph_outline.py
