### 2024-03-05 76ffef3
Remove paper-related doc
