+++ HNX_Paper/paper.bib
@@ -1,229 +1,0 @@
-
-@inproceedings{Joslyn2020DNS,
-	address = {Cham},
-  doi={10.1007/978-3-030-48478-1_1},
-	author = {Joslyn, Cliff A. and Aksoy, Sinan and Arendt, Dustin and Firoz, Jesun and Jenkins, Louis and Praggastis, Brenda and Purvine, Emilie and Zalewski, Marcin},
-	booktitle = {Algorithms and Models for the Web Graph},
-	editor = {Kami{\'{n}}ski, Bogumi{\l} and Pra{\l}at, Pawe{\l} and Szufel, Przemys{\l}aw},
-	isbn = {978-3-030-48478-1},
-	pages = {1--15},
-	publisher = {Springer International Publishing},
-	title = {Hypergraph Analytics of Domain Name System Relationships},
-	year = {2020}}
-
-@article{Molnar2022Application,
-  title={The application of directed hyper-graphs for analysis of models of information systems},
-  doi={10.3390/math10050759},
-  author={Moln{\'a}r, B{\'a}lint and Bencz{\'u}r, Andr{\'a}s},
-  journal={Mathematics},
-  volume={10},
-  number={5},
-  pages={759},
-  year={2022},
-  publisher={MDPI}
-}
-
-/// Tony's work to quickly compute overlap sizes
-@inproceedings{Liu2021Parallel,
-  title={Parallel algorithms for efficient computation of high-order line graphs of hypergraphs},
-  doi={10.1109/hipc53243.2021.00045},
-  author={Liu, Xu T and Firoz, Jesun and Lumsdaine, Andrew and Joslyn, Cliff and Aksoy, Sinan and Praggastis, Brenda and Gebremedhin, Assefaw H},
-  booktitle={2021 IEEE 28th International Conference on High Performance Computing, Data, and Analytics (HiPC)},
-  pages={312--321},
-  year={2021},
-  organization={IEEE}
-}
-
-@article{Zhou2022Topological,
-  title={Topological simplifications of hypergraphs},
-  doi={10.1109/tvcg.2022.3153895},
-  author={Zhou, Youjia and Rathore, Archit and Purvine, Emilie and Wang, Bei},
-  journal={IEEE Transactions on Visualization and Computer Graphics},
-  year={2022},
-  publisher={IEEE}
-}
-
-@article{Aksoy2020Hypernetwork,
-  title={Hypernetwork science via high-order hypergraph walks},
-  doi={10.1140/epjds/s13688-020-00231-0},
-  author={Aksoy, Sinan G and Joslyn, Cliff and Marrero, Carlos Ortiz and Praggastis, Brenda and Purvine, Emilie},
-  journal={EPJ Data Science},
-  volume={9},
-  number={1},
-  pages={16},
-  year={2020},
-  publisher={Springer Berlin Heidelberg}
-}
-
-@article{Pickard2023HAT,
-  title={HAT: Hypergraph analysis toolbox},
-  doi={10.1371/journal.pcbi.1011190},
-  author={Pickard, Joshua and Chen, Can and Salman, Rahmy and Stansbury, Cooper and Kim, Sion and Surana, Amit and Bloch, Anthony and Rajapakse, Indika},
-  journal={PLOS Computational Biology},
-  volume={19},
-  number={6},
-  pages={e1011190},
-  year={2023},
-  publisher={Public Library of Science San Francisco, CA USA}
-}
-
-@misc{Berge1973Graphs,
-  title={Graphs and {H}ypergraphs,(translated by {E}dward {M}inieka)},
-  doi={10.1016/s0924-6509(09)x7013-3},
-  author={Berge, Claude},
-  year={1973},
-  publisher={American Elsevier, New York}
-}
-
[…]
