+++ hypernetx/drawing/rubber_band.py
@@ -146,7 +146,7 @@ def draw_hyper_edge_labels(
         ax.annotate(s, xy, rotation=theta, ha="center", va="center", **params)
 
 
-def layout_hyper_edges(H, pos, node_radius={}, dr=None, contain_hyper_edges=False):
+def layout_hyper_edges(H, pos, node_radius={}, dr=None):
     """
     Draws a convex hull for each edge in H.
 
@@ -191,17 +191,12 @@ def layout_hyper_edges(H, pos, node_radius={}, dr=None, contain_hyper_edges=Fals
     def get_padded_hull(uid, edge):
         # make sure the edge contains at least one node
         if len(edge):
-            points = [
-                cp * (node_radius.get(v, r0) + dr * (2 + radii[v][uid])) + pos[v]
-                for v in edge
-            ]
-
-            if contain_hyper_edges:
-                points.append(cp*r0 + pos[uid])
-
-            points = np.vstack(points)
-
-
+            points = np.vstack(
+                [
+                    cp * (node_radius.get(v, r0) + dr * (2 + radii[v][uid])) + pos[v]
+                    for v in edge
+                ]
+            )
         # if not, draw an empty edge centered around the location of the edge node (in the bipartite graph)
         else:
             points = 4 * r0 * cp + pos[uid]
@@ -210,13 +205,10 @@ def layout_hyper_edges(H, pos, node_radius={}, dr=None, contain_hyper_edges=Fals
 
         return hull.points[hull.vertices]
 
-    return [
-        get_padded_hull(uid, list(H.edges[uid]))
-        for uid in H.edges
-    ]
+    return [get_padded_hull(uid, list(H.edges[uid])) for uid in H.edges]
 
 
-def draw_hyper_edges(H, pos, ax=None, node_radius={}, contain_hyper_edges=False, dr=None, **kwargs):
+def draw_hyper_edges(H, pos, ax=None, node_radius={}, dr=None, **kwargs):
     """
     Draws a convex hull around the nodes contained within each edge in H
 
@@ -240,7 +232,7 @@ def draw_hyper_edges(H, pos, ax=None, node_radius={}, contain_hyper_edges=False,
     PolyCollection
         a Matplotlib PolyCollection that can be further styled
     """
-    points = layout_hyper_edges(H, pos, node_radius=node_radius, dr=dr, contain_hyper_edges=contain_hyper_edges)
+    points = layout_hyper_edges(H, pos, node_radius=node_radius, dr=dr)
 
     polys = PolyCollection(points, **inflate_kwargs(H.edges, kwargs))
 
@@ -368,7 +360,6 @@ def draw(
     node_label_alpha=0.35,
     edge_label_alpha=0.35,
     with_additional_edges=None,
-    contain_hyper_edges=False,
     additional_edges_kwargs={},
     return_pos=False,
 ):
@@ -451,11 +442,6 @@ def draw(
         the transparency (alpha) of the box behind text drawn in the figure for node labels
     edge_label_alpha: float
         the transparency (alpha) of the box behind text drawn in the figure for edge labels
-    with_additional_edges: networkx.Graph
-        ...
-    contain_hyper_edges: bool
-        whether the rubber band shoudl be drawn around the location of the edge in the bipartite graph. This may be invisibile unless "with_additional_edges" contains this information.
-
     """
 
     ax = ax or plt.gca()
@@ -482,22 +468,7 @@ def draw(
     edges_kwargs.setdefault("edgecolors", plt.cm.tab10(np.arange(len(H.edges)) % 10))
     edges_kwargs.setdefault("facecolors", "none")
 
-    polys = draw_hyper_edges(
-        H,
-        pos,
[…]
