+++ hypernetx/classes/factory.py
@@ -23,27 +23,27 @@ def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_c
 # Individual factory methods for property stores
 #-------------------------------------------------------------------------------------------------
 
-def to_property_store_from_dataframe(properties, property_type, 
-                                     edge_col = 'edges', node_col = 'nodes', 
+def to_property_store_from_dataframe(properties, property_type,
+                                     edge_col = 'edges', node_col = 'nodes',
                                      misc_cell_properties = 'misc_properties', misc_properties = 'misc_properties',
                                      weight_prop = None, edge_weight_prop = 'weight', node_weight_prop = 'weight',
                                      cell_weight_col = 'weight',
                                      cell_weights = 1.0, default_edge_weight = 1.0, default_node_weight = 1.0):
-    
-    
-    
+
+
+
     """
-    This function creates a pandas dataframe in the correct format given a 
+    This function creates a pandas dataframe in the correct format given a
     pandas dataframe of either cell, node, or edge properties.
-    
+
     Parameters
     ----------
-    
+
     properties : dataframe
         dataframe of properties for either incidences, edges, or nodes
-    
+
     property_type : str
-        type of property dataframe provided. 
+        type of property dataframe provided.
         Either 'edge_properties', 'node_properties', or 'cell_properties'
 
     edge_col : (optional) str | int, default = 0
@@ -94,35 +94,35 @@ def to_property_store_from_dataframe(properties, property_type,
 
     default_node_weight : (optional) int | float, default = 1
         Used when node weight property is missing or undefined
-        
+
     """
-    
+
     properties_df_dict = {}
-    
-    
+
+
     #create a multi-index with level and id columns.
     if property_type == 'cell_properties':
         incidence_pairs = np.array(properties[[edge_col, node_col]]) #array of incidence pairs to use as UIDs.
         indices = [tuple(incidence_pair) for incidence_pair in incidence_pairs]
         multi_index = pd.MultiIndex.from_tuples(indices, names=['edges', 'nodes'])
         # incidence_df = create_df(datadf, uid_col, weight_col, misc_col)
-    elif property_type == 'edge_properties': 
+    elif property_type == 'edge_properties':
         edge_names = np.array(properties[[edge_col]]) #array of edges to use as UIDs.
         indices = [tuple([edge_name[0]]) for edge_name in edge_names]
         multi_index = pd.MultiIndex.from_tuples(indices, names=['edges'])
         # edge_df = create_df(datadf, uid_col, weight_col, misc_col)
-    elif property_type == 'node_properties': 
+    elif property_type == 'node_properties':
         node_names = np.array(properties[[node_col]]) #array of edges to use as UIDs.
         indices = [tuple([node_name[0]]) for node_name in node_names]
         multi_index = pd.MultiIndex.from_tuples(indices, names=['nodes'])
-        
+
 
     #get names of property columns provided that are not the edge or node columns
     property_columns = set(list(properties.columns)) - set([edge_col, node_col])
     for prop in property_columns:
         properties_df_dict[prop] = properties.loc[:, prop]
-        
-        
+
+
     #get column names if integer was provided instead
     if isinstance(edge_col, int):
         edge_col = properties.columns[edge_col]
@@ -133,14 +133,14 @@ def to_property_store_from_dataframe(properties, property_type,
     if isinstance(node_weight_prop, int):
         node_weight_prop = properties.columns[node_weight_prop]
     if isinstance(misc_properties, int):
-        misc_properties = properties.columns[misc_properties]  
+        misc_properties = properties.columns[misc_properties]
     if isinstance(cell_weight_col, int):
[…]
