### 2024-03-06 fe07fa1
Updating factory to have single dataframe to property store function
