+++ hypernetx/classes/factory.py
@@ -1,21 +1,8 @@
 
+
 import pandas as pd
 import networkx as nx
 import numpy as np
-
-# In[ ]:
-#-------------------------------------------------------------------------------------------------
-# Individual factory methods for incidence stores
-#-------------------------------------------------------------------------------------------------
-
-def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_col = 'edges', node_col = 'nodes'):
-    #change column names to edges and nodes.
-    column_renaming_dict = {edge_col: 'edges', node_col: 'nodes'}
-    IS = pandas_dataframe.rename(column_renaming_dict)
-    return IS
-
-
-
 
 
 # In[ ]:
@@ -23,12 +10,134 @@ def to_incidence_store_from_two_column_pandas_dataframe(pandas_dataframe, edge_c
 # Individual factory methods for property stores
 #-------------------------------------------------------------------------------------------------
 
-def to_property_store_from_dataframe(properties, property_type,
+
+def remove_property_store_duplicates(PS, default_uid_cols, aggregation_methods = {}):
+    for col in PS.columns:
+        if col not in aggregation_methods:
+            aggregation_methods[col] = 'first'
+    return PS.groupby(level = default_uid_cols).agg(aggregation_methods)
+    
+    
+def create_df(properties, uid_cols, indices, multi_index, 
+              default_uid_cols, weight_prop_col, 
+              misc_prop_col, default_weight, aggregation_methods):
+    
+    # initialize a dictionary to be converted to a pandas dataframe.
+    properties_df_dict = {}
+    
+    #get names of property columns provided that are not the edge or node columns
+    property_columns = set(list(properties.columns)) - set(uid_cols)
+    for prop in property_columns: #set those as rows in DF
+        properties_df_dict[prop] = properties.loc[:, prop]
+    
+    
+    #get column names if integer was provided instead and create new uid_cols with string names.
+    uid_cols_to_str = []
+    for col in uid_cols:
+        if isinstance(col, int):
+            uid_cols_to_str.append(properties.columns[col])
+        else:
+            uid_cols_to_str.append(col)
+    uid_cols = uid_cols_to_str
+    if isinstance(weight_prop_col, int):
+        weight_prop_col = properties.columns[weight_prop_col]
+    if isinstance(misc_prop_col, int):
+        misc_prop_col = properties.columns[misc_prop_col]
+        
+        
+    #rename uid columns if needed to default names. No way of doing this correctly without knowing data type.
+    for i in range(len(uid_cols)):    
+        col, default_col = uid_cols[i], default_uid_cols[i]
+        #change name of edges column to "edges"
+        if col != default_col and col in properties_df_dict:
+            # Pop the value associated with col key and store it
+            column_values = properties_df_dict.pop(col)
+            # Add the popped value with the correct col name
+            properties_df_dict[default_col] = column_values
+
+    
+    # set weight column code:
+    # check if weight column exists or if weight col name exists in dictionary and assign it as column if it doesn't
+    # default to use weight column if it exists before looking in misc properties column.
+    if weight_prop_col not in property_columns:
+        create_default_weight_column = False #start by default not setting weight column assuming it exists.
+        if misc_prop_col in property_columns: #check if misc properties exists
+            #make sure that an array of weights wasn't provided in default_weight. use that if it was.
+            if not (isinstance(default_weight, int) or isinstance(default_weight, float)):
+                #check if weight_prop_col is a key in any of the misc properties dicitonary.
+                if any(weight_prop_col in misc_dict for misc_dict in properties_df_dict[misc_prop_col]):
+                    #create list of cell weights from misc properties dictionaries and use default value if not in keys
[…]
