### 2024-03-06 3c56622
returns empty list for nieghbors if level =2
