### 2024-03-06 #6016 @bonicim
@danielskatz I have created a tagged release, [v2.2.0p](https://github.com/pnnl/HyperNetX/releases/tag/v2.2.0p). The rest of the required steps are currently in progress.
