+++ hypernetx/classes/factory.py
@@ -16,21 +16,21 @@ def remove_property_store_duplicates(PS, default_uid_cols, aggregation_methods =
         if col not in aggregation_methods:
             aggregation_methods[col] = 'first'
     return PS.groupby(level = default_uid_cols).agg(aggregation_methods)
-    
-    
-def create_df(properties, uid_cols, indices, multi_index, 
-              default_uid_cols, weight_prop_col, 
+
+
+def create_df(properties, uid_cols, indices, multi_index,
+              default_uid_cols, weight_prop_col,
               misc_prop_col, default_weight, aggregation_methods):
-    
+
     # initialize a dictionary to be converted to a pandas dataframe.
     properties_df_dict = {}
-    
+
     #get names of property columns provided that are not the edge or node columns
     property_columns = set(list(properties.columns)) - set(uid_cols)
     for prop in property_columns: #set those as rows in DF
         properties_df_dict[prop] = properties.loc[:, prop]
-    
-    
+
+
     #get column names if integer was provided instead and create new uid_cols with string names.
     uid_cols_to_str = []
     for col in uid_cols:
@@ -43,10 +43,10 @@ def create_df(properties, uid_cols, indices, multi_index,
         weight_prop_col = properties.columns[weight_prop_col]
     if isinstance(misc_prop_col, int):
         misc_prop_col = properties.columns[misc_prop_col]
-        
-        
+
+
     #rename uid columns if needed to default names. No way of doing this correctly without knowing data type.
-    for i in range(len(uid_cols)):    
+    for i in range(len(uid_cols)):
         col, default_col = uid_cols[i], default_uid_cols[i]
         #change name of edges column to "edges"
         if col != default_col and col in properties_df_dict:
@@ -55,7 +55,7 @@ def create_df(properties, uid_cols, indices, multi_index,
             # Add the popped value with the correct col name
             properties_df_dict[default_col] = column_values
 
-    
+
     # set weight column code:
     # check if weight column exists or if weight col name exists in dictionary and assign it as column if it doesn't
     # default to use weight column if it exists before looking in misc properties column.
@@ -104,8 +104,8 @@ def create_df(properties, uid_cols, indices, multi_index,
         misc_props = properties_df_dict.pop(misc_prop_col)
         # Add the popped value with the new properties name "misc_properties"
         properties_df_dict['misc_properties'] = misc_props
-        
-        
+
+
     #create dataframe from property store dictionary object
     PS = pd.DataFrame(properties_df_dict)
     #set multi index for dataframe
@@ -113,11 +113,11 @@ def create_df(properties, uid_cols, indices, multi_index,
 
     #remove any NaN values or missing values in weight column
     PS['weight'].fillna(default_weight, inplace = True)
-    
-    
+
+
     # remove any duplicate indices and combine using aggregation methods (defaults to 'first' if none provided).
     PS = remove_property_store_duplicates(PS, default_uid_cols, aggregation_methods = aggregation_methods)
-    
+
     #reorder columns to have properties last
     # Get the column names and the specific column
     column_names = list(PS.columns)
@@ -126,17 +126,17 @@ def create_df(properties, uid_cols, indices, multi_index,
     new_order = [col for col in column_names if col != specific_col] + [specific_col]
     # Reorder the dataframe using reindex
     PS = PS.reindex(columns=new_order)
-            
+
     return PS
-        
-        
+
+
 
 def property_store_from_dataframe(properties, property_type,
                                      edge_col = 'edges', node_col = 'nodes',
                                      misc_cell_properties = 'misc_properties', misc_properties = 'misc_properties',
                                      weight_prop = None, edge_weight_prop = 'weight', node_weight_prop = 'weight',
[…]
