### 2024-03-06 3dd9616
updated hypergraph outline to notes from meeting
