### 2024-03-07 #6016 @danielskatz
@brendapraggastis, @bonicim - sorry, but I see two small changes needed in the paper that were in my PR but didn't make it into the changes you made. Please add a space in this line
> In its latest release, HNX 2.0 uses Pandas dataframes[@reback2020pandas;@mckinney-proc-scipy-2010] as its underlying data structure,

so that it becomes
> In its latest release, HNX 2.0 uses Pandas dataframes [@reback2020pandas;@mckinney-proc-scipy-2010] as its underlying data structure,

And similarly, in
> a statistics package as well as a full suite of hypergraph analysis and visualization tools[@Landry2023].

so that it becomes
> a statistics package as well as a full suite of hypergraph analysis and visualization tools [@Landry2023].
