### 2024-03-07 dc47a2a
fixed spacing typos in paper.md
