+++ HNX_Paper/paper.md
@@ -20,39 +20,39 @@ authors:
   orcid: 0000-0003-2466-199X
   affiliation: "1"
 - name: Mark Bonicillo
-  affiliation: "1" 
+  affiliation: "1"
 - name: Cliff Joslyn
   orcid: 0000-0002-5923-5547
   affiliation: "1"
 - name: Emilie Purvine
   orcid: 0000-0003-2069-5594
-  affiliation: "1" 
+  affiliation: "1"
 - name: Madelyn Shapiro
   orcid: 0000-0002-2786-7056
-  affiliation: "1" 
+  affiliation: "1"
 - name: Ji Young Yun
   affiliation: "1"
 affiliations:
 - name: Pacific Northwest National Laboratory, USA
   index: 1
-  
+
 date: 21 June 2023
 bibliography: paper.bib
- 
+
 ---
- 
+
 # Summary
 HyperNetX (HNX) is an open source Python library for the analysis and visualization of complex network data modeled as hypergraphs.
 Initially released in 2019, HNX facilitates exploratory data analysis of complex networks using algebraic topology, combinatorics, and generalized hypergraph and graph theoretical methods on structured data inputs.
 With its 2023 release, the library supports attaching metadata, numerical and categorical, to nodes (vertices) and  hyperedges, as well as to node-hyperedge pairings (incidences).
-HNX has a customizable Matplotlib-based visualization module as well as HypernetX-Widget, its JavaScript addon for interactive exploration and visualization of hypergraphs within Jupyter Notebooks. Both packages are available on GitHub and PyPI. With a growing community of users and collaborators, HNX has become a preeminent tool for hypergraph analysis. 
+HNX has a customizable Matplotlib-based visualization module as well as HypernetX-Widget, its JavaScript addon for interactive exploration and visualization of hypergraphs within Jupyter Notebooks. Both packages are available on GitHub and PyPI. With a growing community of users and collaborators, HNX has become a preeminent tool for hypergraph analysis.
 
 ![HNX-Widget is an add-on for the Jupyter Notebook
 computational environment, enabling users to view and interactively
 explore hypergraphs.
 The main features of the tool are: 1) adjustable layout 2) advanced
 selection and 3) visual encoding of node and edge properties.
-Metadata may be attached to the tool by providing tabular data via two optional data frames indexed by node and hyperedge identifiers. Above is an HNX-Widget visualization of a Scene to Character mapping from the LesMis dataset [@knuth1993].](Figures/hnxexample.png){height="225pt"} 
+Metadata may be attached to the tool by providing tabular data via two optional data frames indexed by node and hyperedge identifiers. Above is an HNX-Widget visualization of a Scene to Character mapping from the LesMis dataset [@knuth1993].](Figures/hnxexample.png){height="225pt"}
 
 
 # Statement of need
@@ -69,24 +69,24 @@ With the development of hypergraph modeling methods, new software was required t
 experimentation and exploration, which prompted the development of HyperNetX.
 
 ## Related Software
-Due to the diversity of hypergraph modeling applications, hypergraph software libraries are 
-often bootstrapped using data structures and methods most appropriate to their usage. 
-In 2020 SimpleHypergraph.jl  was made available for high performance computing on hypergraphs using Julia. 
-The library offers a suite of tools for centrality analysis and community detection and integrates its own 
-visualization tools with those offered by HNX [@Szufel2019]. In 2021 CompleX Group Interactions (XGI)  was released. 
-Originally developed to efficiently discover spreading processes in complex social systems, the library now offers 
-a statistics package as well as a full suite of hypergraph analysis and visualization tools[@Landry2023]. 
-More recently, in 2023 HyperGraphX (HGX)  was released, again with a full suite of tools for community detection 
[…]
