### 2024-03-07 0b5cf2e
updated factory to follow hypergraph_outline structure
