+++ hypernetx/classes/factory.py
@@ -131,16 +131,12 @@ def create_df(properties, uid_cols, indices, multi_index,
 
 
 
-def property_store_from_dataframe(properties, property_type,
-                                     edge_col = 'edges', node_col = 'nodes',
-                                     misc_cell_properties = 'misc_properties', misc_properties = 'misc_properties',
-                                     weight_prop = None, edge_weight_prop = 'weight', node_weight_prop = 'weight',
-                                     cell_weight_col = 'weight',
-                                     cell_weights = 1.0, default_edge_weight = 1.0, default_node_weight = 1.0,
-                                     aggregation_methods = {}):
-
-
-
+def dataframe_factory_method(DF, uid_cols, 
+                             misc_properties_col = 'misc_properties', 
+                             weight_col = 'weight', 
+                             default_weight = 1.0,
+                             aggregate_by = {}):
+    
     """
     This function creates a pandas dataframe in the correct format given a
     pandas dataframe of either cell, node, or edge properties.
@@ -148,64 +144,24 @@ def property_store_from_dataframe(properties, property_type,
     Parameters
     ----------
 
-    properties : dataframe
+    DF : dataframe
         dataframe of properties for either incidences, edges, or nodes
+        
+    uid_cols : list of str or int
+        column index (or name) in pandas.dataframe
+        used for (hyper)edge, node, or incidence (edge, node) IDs. 
 
-    property_type : str
-        type of property dataframe provided.
-        Either 'edge_properties', 'node_properties', or 'cell_properties'
-
-    edge_col : (optional) str | int, default = 0
-        column index (or name) in pandas.dataframe or numpy.ndarray,
-        used for (hyper)edge ids. Will be used to reference edgeids for
-        all set systems.
-
-    node_col : (optional) str | int, default = 1
-        column index (or name) in pandas.dataframe or numpy.ndarray,
-        used for node ids. Will be used to reference nodeids for all set systems.
-
-    cell_weight_col : (optional) str | int, default = None
-        column index (or name) in pandas.dataframe or numpy.ndarray used for
-        referencing cell weights. For a dict of dicts references key in cell
-        property dicts.
-
-    cell_weights : (optional) Sequence[float,int] | int |  float , default = 1.0
-        User specified cell_weights or default cell weight.
-        Sequential values are only used if setsystem is a
-        dataframe or ndarray in which case the sequence must
-        have the same length and order as these objects.
-        Sequential values are ignored for dataframes if cell_weight_col is already
-        a column in the data frame.
-        If cell_weights is assigned a single value
-        then it will be used as default for missing values or when no cell_weight_col
-        is given.
-
-    misc_cell_properties : (optional) str | int, default = None
-        Column name of dataframe corresponding to a column of variable
-        length property dictionaries for the cell. Ignored for other setsystem
-        types.
-
-    misc_properties : (optional) int | str, default = None
+    misc_properties_col : (optional) int | str, default = None
         Column of property dataframes with dtype=dict. Intended for variable
         length property dictionaries for the objects.
 
-    edge_weight_prop : (optional) str, default = None,
+    weight_col : (optional) str, default = None,
         Name of property in edge_properties to use for weight.
 
-    node_weight_prop : (optional) str, default = None,
-        Name of property in node_properties to use for weight.
-
-    weight_prop : (optional) str, default = None
-        Name of property in properties to use for 'weight'
-
[…]
