### 2024-03-07 b43ec9e
Added dictionary factory method and more examples
