+++ hypernetx/classes/factory.py
@@ -12,10 +12,13 @@ import numpy as np
 
 
 def remove_property_store_duplicates(PS, default_uid_cols, aggregation_methods = {}):
+    agg_methods = {}
     for col in PS.columns:
         if col not in aggregation_methods:
-            aggregation_methods[col] = 'first'
-    return PS.groupby(level = default_uid_cols).agg(aggregation_methods)
+            agg_methods[col] = 'first'
+        else:
+            agg_methods[col] = aggregation_methods[col]
+    return PS.groupby(level = default_uid_cols).agg(agg_methods)
 
 
 def create_df(properties, uid_cols, indices, multi_index,
@@ -113,8 +116,8 @@ def create_df(properties, uid_cols, indices, multi_index,
 
     #remove any NaN values or missing values in weight column
     PS['weight'].fillna(default_weight, inplace = True)
-
-
+    
+    
     # remove any duplicate indices and combine using aggregation methods (defaults to 'first' if none provided).
     PS = remove_property_store_duplicates(PS, default_uid_cols, aggregation_methods = aggregation_methods)
 
@@ -131,7 +134,7 @@ def create_df(properties, uid_cols, indices, multi_index,
 
 
 
-def dataframe_factory_method(DF, uid_cols, 
+def dataframe_factory_method(DF, uid_cols, default_uid_cols = None, 
                              misc_properties_col = 'misc_properties', 
                              weight_col = 'weight', 
                              default_weight = 1.0,
@@ -174,17 +177,19 @@ def dataframe_factory_method(DF, uid_cols,
     if len(uid_cols) != 1 and len(uid_cols) != 2:
         raise ValueError("uid_col must be a list and have length of 1 or 2.")
     
-    
-    if DF is None: #if no properties are provided for that property type.
+    #use provided column uids if none are specified
+    if default_uid_cols == None:
         default_uid_cols = uid_cols
-        multi_index = pd.MultiIndex.from_tuples(levels=[[],[]], codes=[[],[]], names=default_uid_cols)
+        
+    if DF is None: #if no properties are provided for that property type.
+        multi_index = pd.MultiIndex.from_tuples([], names=default_uid_cols)
+
         PS = pd.DataFrame(index = multi_index, columns=['weight', 'misc_properties'])
 
     else:
         uids = np.array(DF[uid_cols]) #array of incidence pairs to use as UIDs.
         indices = [tuple(uid) for uid in uids]
-        multi_index = pd.MultiIndex.from_tuples(indices, names=uid_cols)
-        default_uid_cols = uid_cols
+        multi_index = pd.MultiIndex.from_tuples(indices, names=default_uid_cols)
         PS = create_df(DF,
                        uid_cols = uid_cols,
                        default_uid_cols = default_uid_cols,
@@ -208,7 +213,7 @@ def dict_depth(dic, level = 0):
         return level
     return max(dict_depth(dic[key], level + 1) for key in dic) 
 
-def dict_factory_method(D, is_setsystem,
+def dict_factory_method(D, is_setsystem, default_uid_cols = None,
                         misc_properties_col = 'misc_properties', 
                         weight_col = 'weight',
                         default_weight = 1.0,
@@ -222,6 +227,8 @@ def dict_factory_method(D, is_setsystem,
     is_setsystem : bool
         Does the provided dictionary a setsystem (i.e., holds structure)?
         If false then the dictionary should be for edge or node properties.
+    default_uid_cols : TYPE
+        DESCRIPTION.
     misc_properties_col : TYPE
         DESCRIPTION.
     weight_col : TYPE
@@ -237,35 +244,66 @@ def dict_factory_method(D, is_setsystem,
 
     '''
     
+    #use provided column uids if specified
+    
+    if D is None:
+        D = {}
+    
+    
     if is_setsystem:
[…]
