### 2024-03-07 c400777
added list factory method
