### 2024-03-08 4a9752a
Update property store, tests, notebook
