+++ hypernetx/classes/property_store.py
@@ -1,199 +1,56 @@
-from abc import ABC, abstractmethod
 from typing import Any
 from collections.abc import Hashable
 
-import pandas as pd
+from pandas import DataFrame
 
 UUID = "uuid"
+ID = "id"
 WEIGHT = "weight"
 PROPERTIES = "misc_properties"
 
 
-class PropertyStore(ABC):
-    def __init__(self):
-        pass
-    # @property
-    # @abstractmethod
-    # def properties(self) -> Any:
-    #     """Properties assigned to all items in the underlying data table
+class PropertyStore:
+    """Class for storing properties of a collection of edges, nodes, or incidences.
 
-    #     Returns
-    #     -------
-    #     Object containing all properties of the underlying data table
-    #     """
-    #     ...
+    Properties will be stored in a pandas dataframe;
 
-    # @abstractmethod
-    # def get_properties(self, uid: Hashable) -> dict[Any, Any]:
-    #     """Get all properties of an item
-
-    #     Parameters
-    #     ----------
-    #     uid : Hashable
-    #         uid is the index used to fetch all its properties
-
-    #     Returns
-    #     -------
-    #     prop_vals : dict
-    #         ``{named property: property value, ...,
-    #         properties: {property name: property value}}``
-
-
-    #     Raises
-    #     ------
-    #     KeyError
-    #         if (`uid`) is not in :attr:`properties`,
-
-    #     See Also
-    #     --------
-    #     get_property, set_property
-    #     """
-    #     ...
-
-    # @abstractmethod
-    # def get_property(self, uid: Hashable, prop_name: str | int) -> Any:
-    #     """Get a property of an item
-
-    #     Parameters
-    #     ----------
-    #     uid : Hashable
-    #         uid is the index used to fetch its property
-    #     prop_name : str | int
-    #         name of the property to get
-
-    #     Returns
-    #     -------
-    #     prop_val : any
-    #         value of the property
-
-    #     None
-    #         if property not found
-
-    #     Raises
-    #     ------
-    #     KeyError
-    #         if (`uid`) is not in :attr:`properties`,
-
-    #     See Also
-    #     --------
-    #     get_properties, set_property
-    #     """
-    #     ...
-
-    # @abstractmethod
-    # def set_property(self, uid: Hashable, prop_name: str | int, prop_val: Any) -> None:
-    #     """Set a property of an item in the 'properties' colelction
-
-    #     Parameters
-    #     ----------
-    #     uid : Hashable
-    #         uid is the index used to set its property
-    #     prop_name : str | int
-    #         name of the property to set
-    #     prop_val : any
-    #         value of the property to set
-
-    #     Raises
-    #     ------
-    #     KeyError
-    #         If (`uid`) is not in :attr:`properties`
-
-
-    #     See Also
-    #     --------
-    #     get_property, get_properties
-    #     """
-
-    # def __iter__(self):
-    #     """
-    #     Returns an iterator object for iterating over data in PropertyStore
-
-    #     Returns:
-    #     -------
-    #     DataIterator:
-    #         An iterator object
-    #     """
-    #     return self
-
-    # def __iter__(self):
-    #     """Returns an iterator over items in the underlying data table
-
-    #     Returns
-    #     -------
-    #     Iterator
-
-    #     """
-    #     ...
-
-    # def __len__(self) -> int:
-    #     """Number of items in the underlying data table
-
-    #     Returns
-    #     -------
-    #     int
-    #     """
-    #     ...
-
-    # def __getitem__(self, uid) -> dict:
-    #     """Returns the common attributes (e.g. weight) and properties of an item in the underlying data table
-
-    #     Returns same result as get_property
-
-    #     Parameters
-    #     ----------
-    #     Hashable
-    #         uid is the index used to set its property
-
[…]
