### 2024-03-09 #6016 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
